### 2024-06-21 #6911 @logological
@crvernon: I need to take most of the rest of this month off as I'm moving. If no other editor accepts by July 2, I'm happy to take it on then.
