### 2024-06-22 #6911 @logological
@haider8645 I note that your manuscript relies heavily on figures, none of which seem to have been successfully embedded into your PDF, and only a few of which are actually referred to in the text.  I'm hesitant to contact any potential reviewers without having a reasonably complete manuscript to show them.  Could you please address this issue as a matter of priority?  Please refer to the ["Figures and images"](https://joss.readthedocs.io/en/latest/paper.html#figures-and-images) and ["Tables and figures"](https://joss.readthedocs.io/en/latest/paper.html#tables-and-figures) sections of the JOSS Paper Format documentation for instructions on including and referring to images in your paper.  Please let me know once you've solved the issue, or if you require further assistance.
