### 2024-06-24 #6911 @haider8645
> @editorialbot add @logological as editor
>
> Thanks @logological for the quick response! No problem about the delay as all of our other editors are tied up right now. I am going to go ahead and assign you to this one to get it on the books but feel free to hold any action on your part until July 2. Good luck with your move!
>
> In the meantime @haider8645, please direct any questions you may have before July 2 to me. Some things you can address while we are waiting to get started:
>
>     * Recommend a few reviewers that you think could provide a review of your submission.
>
>     * Get your paper length down to around 1000 words.
>
>     * Fix how your citations an bibliography are used, this is currently incorrect in your paper.  You can check out the submission docs for help on this and other recently published JOSS articles.
>
>
> Thanks!
Thanks @crvernon for the feeback! I will get to it to fix the changes you mentioned.
