### 2024-06-24 c18b177
Fix citation of references and visuals of figures in paper
