+++ paper/paper.md
@@ -25,24 +25,20 @@ bibliography: paper.bib
 # Summary
 Navigation of autonomous mobile robots is a complex problem which comprises of multiple software components, each performing a specific task, which when combined form the navigation software architecture. The ugv_nav4d is a path planner which can be used to navigate complex indoor and outdoor environments. It provides a traversable, motion constraint compliant, and collision free trajectory. It is highly customizable and can be used for any type of terrestrial robot.
 
-To find a traversable trajectory, ugv_nav4d internally maintains a traversability map (TraversabilityMap3d) of the environment. The TraversabilityMap3d [1] is generated based on a multi-layered surface map (MLSMap [1]), which enables ugv_nav4d to plan trajectories in Multi-surface environments. The trajectory is composed by careful selection and piecing together of motion primitives from a pool of primitives. The primitives are generated based on the specific mechnical features of the robot. As a result, the planned trajectory always fits well with the motion capabilites of the robot. 
+To find a traversable trajectory, ugv_nav4d internally maintains a traversability map (TraversabilityMap3d) of the environment. The TraversabilityMap3d [@slammaps] is generated based on a multi-layered surface map (MLSMap [@slammaps]), which enables ugv_nav4d to plan trajectories in multi-surface environments. The trajectory is composed by careful selection and piecing together of motion primitives from a pool of primitives. The primitives are generated based on the specific mechnical features of the robot. As a result, the planned trajectory always fits well with the motion capabilites of the robot. 
 
 In short, ugv_nav4d
 
-- can plan continuous, smooth, and feasible paths in complex indoor and outdoor environments
-- can plan in Multi-surface environments, e.g. natural cave system, multi-storey buildings
-- provides a complete trajectory for the planned path which is ready for execution
-- provides a graphical user interface (GUI) for introspection and debugging
+- can plan continuous, smooth, and feasible paths in complex indoor and outdoor environments.
+- can plan in multi-surface environments, e.g. natural cave system, multi-storey buildings.
+- provides a complete trajectory for the planned path which is ready for execution.
+- provides a graphical user interface (GUI) for introspection and debugging.
 
-<figure>
-  <img src="figures/parking_deck.png" alt="Figure 1" width="600"/>
-  <figcaption>Figure 1: Planned trajectory in a multi-storey environment.</figcaption>
-</figure>
-
+![Planned trajectory in a multi-storey environment.](figures/parking_deck.png){width="300pt"}
 
 # Statement of need
 - General purpose path planner is needed which can support any type of terrestrial mobile robot system e.g. Differential Drive, Omni-directional, Ackermann Drive etc. 
-- Capability to plan in a Multi-surface unstructured (e.g. natural cave system) and structured (e.g. office, garage) environments
+- Capability to plan in a multi-surface unstructured (e.g. natural cave system) and structured (e.g. office, garage) environments.
 
 # Software Components 
 The core software components are 
@@ -52,12 +48,9 @@ The core software components are
 - PreComputedMotions
 
 ### EnvironmentXYZTheta 
[…]
