+++ paper/paper.bib
@@ -17,32 +17,37 @@
   author = {Janosch Machowinski and Arne Böckmann},
   title = {Traversability Generator3d},
   howpublished = {https://github.com/dfki-ric/traversability_generator3d.git},
-  note = {Accessed: June 20, 2024}
+  note = {Accessed: June 20, 2024},
+  year = {2016}
 }
 
 @misc{roboticsElab,
   title = {Robotics Exploration Laboratory},
   url = {https://www.dfki.de/web/technologien-anwendungen/living-labs/robotics-exploration-laboratory},
-  note = {Accessed: June 20, 2024}
+  note = {Accessed: June 20, 2024},
+  year = {2024}
 }
 
 @software{trajectoryfollower,
   author = {Ajish Babu},
   title = {Trajectory Follower},
   url = {https://github.com/rock-control/control-trajectory_follower},
-  note = {Accessed: June 20, 2024}
+  note = {Accessed: June 20, 2024},
+  year = {2016}
 }
 
 @software{sbplsplineprimitives,
   author = {Arne Böckmann},
   title = {SBPL Spline Primitives},
   url = {https://github.com/rock-planning/planning-sbpl_spline_primitives},
-  note = {Accessed: June 20, 2024}
+  note = {Accessed: June 20, 2024},
+  year = {2016}
 }
 
 @software{googletest,
   author = {Google},
   title = {GoogleTest},
   url = {https://github.com/google/googletest},
-  note = {Accessed: June 20, 2024}
+  note = {Accessed: June 20, 2024},
+  year = {2024}
 }
