### 2024-06-24 #6911 @haider8645
@logological Thanks for your feedback. I have fixed the figures and how they are references in parts of the article.
