### 2024-07-05 #6911 @logological
@soraxas @KanishAnand @OlgerSiebinga Would you be willing and able to review this submission on navigation of autonomous robots for the Journal of Open Source Software?
