### 2024-07-06 #6911 @soraxas
Hi @logological yes the paper and repo seems relevant to me, I can review it later.
