### 2024-07-12 #6911 @logological
@brean Would you be willing and able to review this submission on navigation of autonomous robots for the Journal of Open Source Software?
