### 2024-07-14 #6983 @editorialbot
[REVIEW]: Advanced Multi-Surface Navigation for Unmanned Ground Vehicles (UGVs) Using 4D Path Planning Techniques (ugv_nav4d)
**Submitting author:** @haider8645 (<a href="http://orcid.org/0009-0008-1199-3489">Muhammad Haider Khan Lodhi</a>)
**Repository:** https://github.com/dfki-ric/ugv_nav4d
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.1
