### 2024-07-14 #6983 @brean
@logological as I work for the same research institute as the authors of this publication (even in the same department) I think this is clearly a COI and I am not allowed to review this paper, as this situation "being employed at the same institution" is explicitly stated as a COI in the [JOSS Conflict of Interest Policy](https://github.com/openjournals/joss/blob/main/COI.md)
