### 2024-07-15 #6983 @soraxas
Hi @haider8645 I had a quick glance into the paper, and the statement of need doesn't really address why this package is needed in the community.

i.e., how does this differ from other similar-purpose packages like [Nav2](https://nav2.org/) ([github](https://github.com/ros-navigation/navigation2)) [1], one of the most well-structured / generic navigation frameworks that are widely adopted in autonomous robots and tested in the industry (De facto standard in ROS); or other more speciality framework like [vox_nav](https://github.com/NMBURobotics/vox_nav) [2] which is created for unstructured treeains.

The traversability aspect should be able to achieved via imposing cost in like a costmap?

[1] Macenski, S., Martín, F., White, R., & Clavero, J. G. (2020, October). The marathon 2: A navigation system. In 2020 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS) (pp. 2718-2725). IEEE.
[2] Atas, F., Cielniak, G., & Grimstad, L. (2022, October). Elevation state-space: Surfel-based navigation in uneven environments for mobile robots. In 2022 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS) (pp. 5715-5721). IEEE.
