### 2024-07-18 #6983 @soraxas
> @soraxas Excuse me for the late reply. I am on vacation atm so my replies might be delayed. Hmmm, please see some brief comments below about nav2 for now. If you expect a more detailed analysis of the capabilities then let me know but it would just take me a bit longer.
>
> 1. The environment representation used by the planner to generate the Traversability map is based on a Multi-Layered Surface Map (MLS). The MLS map can be generated online or an offline MLS map can be used. As of my latest update, ROS2 does not natively support MLS (Multi-Layered Surface) maps out-of-the-box within the standard Nav2 stack without additional development or customization. The primary map type supported by Nav2 is the 2D occupancy grid map. On the contrary, we used a 3D Map. The choice of MLS maps vs Voxel maps (used in nav2) is a personal choice. I personally like MLS Maps because they are great for visualization and provide a detailed surface representation.
> 2. The key contribution is that the planner can plan in a multistorey environment without the need to maintain multiple costmaps for each floor of a building. The planner internally maintains a 3D Traversablity Map for long distance planning in environments with distinct elevation levels, such as buildings with multiple floors or terrains with varying heights. To the best of my knowledge, ros2 nav2 does not natively support multistorey navigation out-of-the-box.
> 3. The Traversability map servers a similar purpose as the costmap in ROS but each cell in the traversability map contains further information about the ground surface like slope, plane model etc.  See [TravGenTrackingData](https://github.com/dfki-ric/traversability_generator3d/blob/471242ea9079428037e47ab9bd1ecce7f01b3ec8/src/TravGenNode.hpp#L11). This can be seen as kind of an alternative to the combination of 3d costmap + gradient map in nav2. See [Other Forms](https://docs.nav2.org/concepts/index.html#environmental-representation)
>
> So in short, the ugv_nav4d provides many new capabilities which are not yet available in nav2. We are currently working on making a ROS2 wrapper for the planner so that the ROS2 community can used it out-of-the-box. A fully integrated plugin support for nav2 is not in the pipeline but can be done if the interest from the community side is high :)
>
> @arneboe FYI

Thanks for the response. I would suggest adding snippets of these relevant disussion to the statement of need section. I will move our conversion to issues in the target repository.
