### 2024-07-18 dfki-ric/ugv_nav4d#3 @soraxas
Tracking PLY file in repo
Should the PLY files be tracked inside this source repo?

It contains > 100 MB just to clone the repo.

```sh
$ du -h -d1
36M	./.git
252K	./src
3.0M	./doc
2.0M	./paper
28K	./source_dependencies
119M	./test_data
160M	.
```

(part of https://github.com/openjournals/joss-reviews/issues/6983)
