### 2024-07-18 dfki-ric/ugv_nav4d#5 @soraxas
Add system requirements to Docunments/Readme
The `CMakeLists.txt` calls scripts that assumes the host computer to be on Ubuntu-like system; however, this system requirements is not stated anywhere.

https://github.com/dfki-ric/ugv_nav4d/blob/86650761370d2ed10e6fc49aa6ff797c18182a13/source_dependencies/install_os_dependencies.bash#L1-L7

(part of https://github.com/openjournals/joss-reviews/issues/6983)
