### 2024-07-18 dfki-ric/ugv_nav4d#6 @soraxas
Build system depends on `Rock` 
It seems the build system depends on `rock`, a build system macro developed by `dfki.de`:

https://github.com/dfki-ric/ugv_nav4d/blob/86650761370d2ed10e6fc49aa6ff797c18182a13/CMakeLists.txt#L28-L31

However, its not part of the install target in the `source_dependencies` scripts.

The Readme file contains the following text:

```
Compiling inside a ROCK environment
See the manifest.xml for an up to date list of dependencies.

Buidling inside ROCK works as usual. The package can be found in the backbone package set: https://git.hb.dfki.de/sw-backbone/package_sets
```

However, the gitlab link https://git.hb.dfki.de/sw-backbone/package_sets is private and requires login?

(part of https://github.com/openjournals/joss-reviews/issues/6983)
