### 2024-07-18 dfki-ric/ugv_nav4d#7 @soraxas
Community guidelines
Needs the following as part of the documentation

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

(part of https://github.com/openjournals/joss-reviews/issues/6983)
