### 2024-07-18 dfki-ric/ugv_nav4d#8 @soraxas
Unit Tests and CI
> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

In https://github.com/dfki-ric/ugv_nav4d/blob/main/src/test/test_ugv_nav4d.cpp there contains a few google test function.
But I wouldn't call it unit testing as it's only setting up 1 scenario and tested:
- setting up grid resolution
- valid / invalid start/end states
- tested it can finishes (finding solution) for 1 particular planning request

Can we have some more structured unit testing in the traditional sense?

(part of https://github.com/openjournals/joss-reviews/issues/6983)
