### 2024-07-18 dfki-ric/ugv_nav4d#8 @soraxas
The CI file in https://github.com/dfki-ric/ugv_nav4d/blob/main/.gitlab-ci.yml is using Gitlab format.

Presumably, it's from `dfki-ric`'s private gitlab instance? If this public instance will be in Github, you should use github action instead

https://docs.github.com/en/actions
