### 2024-07-20 #6983 @logological
@AtsushiSakai @mjansen4857 @whoenig Would you be willing and able to review this submission on navigation of autonomous robots for the Journal of Open Source Software?
