### 2024-07-31 dfki-ric/ugv_nav4d#2 @planthaber
The MLS approach has no discretization in the z axis. The x,y coordinates are formed by a grid with a cell size, but the z axis is not discretized. The cell content is a list of blocks with a start and stop height (float). Additionally the top of those blocks can have a slope value.

This means the MLS has a smooth surface, which is better suited for **uneven** ground navigation than occupancy voxels that create steps on sloped surfaces.

The original idea of the MLS is described here: https://cvg.cit.tum.de/_media/spezial/bib/triebel06multi.pdf

Voxel maps are imo better suited for drones or underwater vehicles, where non-ground obstacles have more importance, but when planning for ground vehicles in uneven terrain that addional accuracy makes the difference.
