### 2024-07-31 193d085
Add community guidelines and update readme with system requirements
