### 2024-07-31 dfki-ric/ugv_nav4d#7 @haider8645
Added a CONTRIBUTING.md file to the repo
