### 2024-07-31 dfki-ric/ugv_nav4d#5 @haider8645
Added the system requirements with a clear mention of the ubuntu requirement as a base OS.
