### 2024-07-31 dfki-ric/ugv_nav4d#6 @haider8645
We use the library as part of the ROCK robotics framework developed and used at DFKI. However, from the open source user community view, we use this [CMakeLists](https://github.com/dfki-ric/ugv_nav4d/blob/main/source_dependencies/CMakeLists.txt) which together with installation steps should automatically setup and install all the dependencies.
