### 2024-07-31 dfki-ric/ugv_nav4d#3 @haider8645
Good point, I guess it would be better to move the ply file(s) to a external storage location. Do you have any recommendation or am I free to choose any?
