### 2024-07-31 fee62cd
Add API documentation instructions and cleanup unused figures
