+++ paper/paper.md
@@ -23,9 +23,9 @@ bibliography: paper.bib
 ---
 
 # Summary
-Navigation of autonomous mobile robots is a complex problem which comprises of multiple software components, each performing a specific task, which when combined form the navigation software architecture. The ugv_nav4d is a path planner which can be used to navigate complex indoor and outdoor environments. It provides a traversable, motion constraint compliant, and collision free trajectory. It is highly customizable and can be used for any type of terrestrial robot.
+Navigation of autonomous mobile robots is a complex problem which comprises of multiple software components, each performing a specific task, which when combined form the navigation software architecture. Ugv_nav4d can be used to navigate complex indoor and outdoor environments. It provides a traversable, motion constraint compliant, and collision free trajectory. It is highly customizable and can be used for any type of terrestrial robot.
 
-To find a traversable trajectory, ugv_nav4d internally maintains a traversability map (TraversabilityMap3d) of the environment. The TraversabilityMap3d [@slammaps] is generated based on a multi-layered surface map (MLSMap [@slammaps]), which enables ugv_nav4d to plan trajectories in multi-surface environments. The trajectory is composed by careful selection and piecing together of motion primitives from a pool of primitives. The primitives are generated based on the specific mechnical features of the robot. As a result, the planned trajectory always fits well with the motion capabilites of the robot. 
+To find a traversable trajectory, ugv_nav4d internally maintains a 3D Traversability Map (TraversabilityMap3d) of the environment. The TraversabilityMap3d [@slammaps] is generated based on a multi-layered surface map (MLS [@mlsmaps], [@slammaps]), which enables ugv_nav4d to plan trajectories in multi-surface environments. The trajectory is composed by careful selection and piecing together of motion primitives from a pool of primitives. The primitives are generated based on the specific mechnical features of the robot. As a result, the planned trajectory always fits well with the motion capabilites of the robot. 
 
 In short, ugv_nav4d
 
@@ -37,9 +37,29 @@ In short, ugv_nav4d
 ![Planned trajectory in a multi-storey environment.](figures/parking_deck.png){width="300pt"}
 
 # Statement of need
-- General purpose path planner is needed which can support any type of terrestrial mobile robot system e.g. Differential Drive, Omni-directional, Ackermann Drive etc. 
-- Capability to plan in a multi-surface unstructured (e.g. natural cave system) and structured (e.g. office, garage) environments.
 
+The ROS navigation stack [@ros2] is undoubtedly the most used open-source robotics framework. This claim can be backed by its widespread usage in research and industry. Navigation2 (nav2) [@macenski2020nav2] is the newest navigation system which is based on the legacy of ROS Navigation. In this paper, we propose a new, fully open-source robot navigation library called ugv_nav4d. Ugv_nav4d presents significant improvements over nav2, particularly in environmental representation and navigation in complex, uneven terrains and multistorey environments. By utilizing Multi-Layered Surface Maps [@mlsmaps] and a 3D Traversability Map [@slammaps], ugv_nav4d offers a more detailed and accurate navigation solution for ground vehicles. These capabilities are essential for operations in challenging terrains and address critical needs within the robotics community. The ongoing development of a ROS2 wrapper for ugv_nav4d aims to facilitate its adoption and integration within the ROS2 ecosystem, providing substantial benefits to the community and enhancing the capabilities of ground vehicle navigation.
+
+### Environment Representation and Mapping Techniques
[…]
