+++ paper/paper.bib
@@ -51,3 +51,45 @@
   note = {Accessed: June 20, 2024},
   year = {2024}
 }
+
+@INPROCEEDINGS{mlsmaps,
+  author={Triebel, Rudolph and Pfaff, Patrick and Burgard, Wolfram},
+  booktitle={2006 IEEE/RSJ International Conference on Intelligent Robots and Systems}, 
+  title={Multi-Level Surface Maps for Outdoor Terrain Mapping and Loop Closing}, 
+  year={2006},
+  volume={},
+  number={},
+  pages={2276-2282},
+  keywords={Terrain mapping;Mobile robots;Data structures;Path planning;Multilevel systems;Bridges;Buildings;Impedance matching;Robot sensing systems;Large-scale systems},
+  doi={10.1109/IROS.2006.282632}
+}
+
+@article{macenski2020nav2,
+  author       = {Steve Macenski and
+                  Francisco Mart{\'{\i}}n and
+                  Ruffin White and
+                  Jonatan Gin{\'{e}}s Clavero},
+  title        = {The Marathon 2: {A} Navigation System},
+  journal      = {CoRR},
+  volume       = {abs/2003.00368},
+  year         = {2020},
+  url          = {https://arxiv.org/abs/2003.00368},
+  eprinttype    = {arXiv},
+  eprint       = {2003.00368},
+  timestamp    = {Wed, 11 Mar 2020 09:34:48 +0100},
+  biburl       = {https://dblp.org/rec/journals/corr/abs-2003-00368.bib},
+  bibsource    = {dblp computer science bibliography, https://dblp.org}
+}
+
+@article{ros2,
+   title={Robot Operating System 2: Design, architecture, and uses in the wild},
+   volume={7},
+   ISSN={2470-9476},
+   url={http://dx.doi.org/10.1126/scirobotics.abm6074},
+   DOI={10.1126/scirobotics.abm6074},
+   number={66},
+   journal={Science Robotics},
+   publisher={American Association for the Advancement of Science (AAAS)},
+   author={Macenski, Steven and Foote, Tully and Gerkey, Brian and Lalancette, Chris and Woodall, William},
+   year={2022},
+   month=may }
