### 2024-08-01 dfki-ric/ugv_nav4d#2 @haider8645
@soraxas please find the updated paper [here](https://raw.githubusercontent.com/openjournals/joss-papers/joss.06983/joss.06983/10.21105.joss.06983.pdf)
