### 2024-08-01 80bc4bb
Add path to further os dependencies
