+++ .github/workflows/c-cpp.yml
@@ -11,15 +11,8 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v4
-    - name: install_deps
-      run: bash ./source_dependencies/install_os_dependencies.bash
-    - name: build_deps
-      run: |
-        cd source_dependencies
-        bash ./build.bash ../build/install
     - name: build_ugv_nav4d
       run: |
-        cd build
-        source install/env.sh
-        cmake -DCMAKE_INSTALL_PREFIX=./install -DTESTS_ENABLED=OFF -DENABLE_DEBUG_DRAWINGS=OFF -DCMAKE_BUILD_TYPE=RELEASE ..
+        mkdir build && cd build
+        cmake -DINSTALL_DEPS=ON -DCMAKE_INSTALL_PREFIX=./install ..
         make -j install
