### 2024-08-01 dfki-ric/ugv_nav4d#8 @haider8645
Build CI: [Added github actions based build CI](https://github.com/dfki-ric/ugv_nav4d/actions)

Unit Testing: The current tests test basic functionality of the planner. If you want something else then can you point me to an example implementation on another library? Thanks!
