### 2024-08-12 9383e8a
Make vizkit3d_debug_drawings optional
