### 2024-08-15 808bae3
Update cmake for the gui
