### 2024-08-15 e18acb1
Use ENABLE_DEBUG_DRAWINGS instead of ENABLE_V3DD_DRAWINGS
