+++ src/EnvironmentXYZTheta.cpp
@@ -16,8 +16,6 @@ using namespace sbpl_spline_primitives;
 using trajectory_follower::SubTrajectory;
 using trajectory_follower::DriveMode;
 
-//#define ENABLE_V3DD_DRAWINGS
-
 namespace ugv_nav4d
 {
 
@@ -202,14 +200,14 @@ bool EnvironmentXYZTheta::obstacleCheck(const maps::grid::Vector3d& pos, double 
 
         if(stats.getRobotStats().getNumObstacles() || stats.getRobotStats().getNumFrontiers()) 
         {
-    #ifdef ENABLE_V3DD_DRAWINGS
+#ifdef ENABLE_DEBUG_DRAWINGS
             V3DD::COMPLEX_DRAWING([&]()
             {
                 const std::string drawName("ugv_nav4d_obs_check_fail_" + nodeName);
                 V3DD::CLEAR_DRAWING(drawName);
                 V3DD::DRAW_WIREFRAME_BOX(drawName, pos, Eigen::Quaterniond(Eigen::AngleAxisd(discTheta.getRadian(), Eigen::Vector3d::UnitZ())), Eigen::Vector3d(travConf.robotSizeX, travConf.robotSizeY, travConf.robotHeight), V3DD::Color::red);
             });
-    #endif
+#endif
 
             LOG_INFO_S << "Num obstacles: " << stats.getRobotStats().getNumObstacles();
             LOG_INFO_S << "Error: " << nodeName << " inside obstacle";
@@ -225,7 +223,7 @@ bool EnvironmentXYZTheta::checkStartGoalNode(const string& name, traversability_
 
     maps::grid::Vector3d nodePos;
     travGen.getTraversabilityMap().fromGrid(node->getIndex(), nodePos, node->getHeight(), false);
-#ifdef ENABLE_V3DD_DRAWINGS
+#ifdef ENABLE_DEBUG_DRAWINGS
         V3DD::COMPLEX_DRAWING([&]()
         {
             const std::string drawName("ugv_nav4d_check_start_goal_" + name);
@@ -242,7 +240,7 @@ bool EnvironmentXYZTheta::checkStartGoalNode(const string& name, traversability_
 void EnvironmentXYZTheta::setGoal(const Eigen::Vector3d& goalPos, double theta)
 {
 
-#ifdef ENABLE_V3DD_DRAWINGS
+#ifdef ENABLE_DEBUG_DRAWINGS
     V3DD::CLEAR_DRAWING("ugv_nav4d_env_goalPos");
     V3DD::DRAW_ARROW("ugv_nav4d_env_goalPos", goalPos, base::Quaterniond(Eigen::AngleAxisd(M_PI, base::Vector3d::UnitX())),
             base::Vector3d(1,1,1), V3DD::Color::red);
@@ -287,7 +285,7 @@ void EnvironmentXYZTheta::setGoal(const Eigen::Vector3d& goalPos, double theta)
     precomputeCost();
     LOG_INFO_S << "Heuristic computed";
     //draw greedy path
-#ifdef ENABLE_V3DD_DRAWINGS
+#ifdef ENABLE_DEBUG_DRAWINGS
     V3DD::COMPLEX_DRAWING([&]()
     {
         V3DD::CLEAR_DRAWING("ugv_nav4d_greedyPath");
@@ -323,7 +321,7 @@ void EnvironmentXYZTheta::setGoal(const Eigen::Vector3d& goalPos, double theta)
 
 void EnvironmentXYZTheta::expandMap(const std::vector<Eigen::Vector3d>& positions)
 {
-#ifdef ENABLE_V3DD_DRAWINGS
[…]
