+++ src/gui/PlannerGui.cpp
@@ -22,8 +22,6 @@
 
 using namespace ugv_nav4d;
 
-//#define ENABLE_V3DD_DRAWINGS
-
 PlannerGui::PlannerGui(const std::string& dumpName): QObject()
 {
     setupUI();
@@ -66,7 +64,7 @@ void PlannerGui::setupUI()
     goal.orientation.setIdentity();
     
     widget = new vizkit3d::Vizkit3DWidget();
-#ifdef ENABLE_V3DD_DRAWINGS
+#ifdef ENABLE_DEBUG_DRAWINGS
     V3DD::CONFIGURE_DEBUG_DRAWINGS_USE_EXISTING_WIDGET(widget);
 #endif
     trav3dViz.setPluginName("TravMap");
@@ -452,7 +450,7 @@ void PlannerGui::picked(float x, float y, float z, int buttonMask, int modifierM
             start.position << x, y, z;
             start.position.z() += travConfig.distToGround; //because we click on the ground but need to put robot position
 
-#ifdef ENABLE_V3DD_DRAWINGS
+#ifdef ENABLE_DEBUG_DRAWINGS
             V3DD::CLEAR_DRAWING("ugv_nav4d_start_aabb");
             V3DD::DRAW_WIREFRAME_BOX("ugv_nav4d_start_aabb", start.position +  base::Vector3d(0, 0, travConfig.distToGround / 2.0), start.orientation,
                                base::Vector3d(travConfig.robotSizeX, travConfig.robotSizeY, travConfig.robotHeight - travConfig.distToGround), V3DD::Color::cyan);
@@ -540,7 +538,7 @@ void PlannerGui::startOrientationChanged(int newValue)
     const double rad = newValue/180.0 * M_PI;
     start.orientation = Eigen::AngleAxisd(rad, Eigen::Vector3d::UnitZ());
     startViz.setRotation(QQuaternion(start.orientation.w(), start.orientation.x(), start.orientation.y(), start.orientation.z()));
-#ifdef ENABLE_V3DD_DRAWINGS
+#ifdef ENABLE_DEBUG_DRAWINGS
     V3DD::CLEAR_DRAWING("ugv_nav4d_start_aabb");
     V3DD::DRAW_WIREFRAME_BOX("ugv_nav4d_start_aabb", start.position + Eigen::Vector3d(0, 0, travConfig.distToGround),
                        start.orientation, base::Vector3d(travConfig.robotSizeX, travConfig.robotSizeY, travConfig.robotHeight), V3DD::Color::cyan);
@@ -574,7 +572,7 @@ void PlannerGui::startPlanThread()
 {
     bar->setMaximum(0);
     std::thread t([this](){
-#ifdef ENABLE_V3DD_DRAWINGS
+#ifdef ENABLE_DEBUG_DRAWINGS
         V3DD::CONFIGURE_DEBUG_DRAWINGS_USE_EXISTING_WIDGET(this->widget);
 #endif
         this->plan(this->start, this->goal);
