### 2024-08-22 #6983 @OlgerSiebinga
Hi @haider8645,

I have some questions about the authorship and contributions, so I hope you can clarify some things.

First, I would like to ask you about the authorship. I noticed that there are quite some contributors to the repo, some of which have made [substantial contributions](https://github.com/dfki-ric/ugv_nav4d/graphs/contributors) (in terms of lines of code/commits). However, most are not co-authors and are not mentioned in the acknowledgments. According to the [JOSS policy](https://joss.readthedocs.io/en/latest/submitting.html#authorship), it's up to you to decide whom to include. However, I am curious why some major contributors are not mentioned in the paper. Furthermore, I can't find the second author ( Janosch Machowinski) in the list of contributors to the repo. Could you clarify his contribution?

Second, I read in the paper that you're currently working on a ROS2 wrapper. What is the status of that development? I'm asking because the potential impact of your software can be substantially increased by providing this wrapper. I would be more inclined to use your planner if I could directly incorporate it into my robots by accessing it from ROS. Including this wrapper in this published version of the software and paper might be worth considering.
