### 2024-08-22 dfki-ric/ugv_nav4d#6 @OlgerSiebinga
If I try to use the automatic installation, I'm still getting an error that ROCK cannot be installed:

`CMake Error at /usr/share/cmake-3.22/Modules/FindPkgConfig.cmake:603 (message):
  A required package was not found
Call Stack (most recent call first):
  /usr/share/cmake-3.22/Modules/FindPkgConfig.cmake:825 (_pkg_check_modules_internal)
  build/install/share/rock/cmake/Rock.cmake:322 (pkg_check_modules)
  build/install/share/rock/cmake/Rock.cmake:520 (rock_find_pkgconfig)
  build/install/share/rock/cmake/Rock.cmake:944 (rock_target_definition)
  build/install/share/rock/cmake/Rock.cmake:1042 (rock_library_common)
  src/CMakeLists.txt:55 (rock_library)
`

I followed the steps in the README file for automatic installation:

` cd ugv_nav4d
   mkdir build && cd build
   cmake -DINSTALL_DEPS=ON -DCMAKE_INSTALL_PREFIX=./install ..
   make -j install
`
