### 2024-08-22 #6983 @OlgerSiebinga
I have some opportunities to improve the paper itself, listed below in order of appearance. Let me know if I can help by clarifying or providing more explicit examples.

Summary
- the first sentence is a bit confusing to me. what are the specific components and tasks you are talking about? Does ugv_nav4d address all these tasks or just one of them?
statement of need
- A clear problem definition is missing in my opinion. It is there implicitly (in the advantages of ugv_nav4d), but it is not explicitly stated.
- the target audience (and target use case) are not explicitly described
Figures
- I think combining some figures into single figures with multiple subfigures could improve the layout and thereby the readability of the paper (e.g., figures 3 and 4)
Quality of writing
- The paper could benefit from some editing for structure. specifically:
 	o Not all paragraphs logically follow the previous paragraph
	o The summary contains some duplicate information
	o Some of the lists of bullet points would be more convincing when written out in full text
	o The second half of the "statement of need" seems to deal more with the inner workings of the software than the actual need for the software.
	o The word count of the paper is over 2000 while it should be between 250 and 1000 according to the [read the docs page](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain)
