### 2024-08-30 dfki-ric/ugv_nav4d#6 @haider8645
Hmm, that is very strange. I followed the same steps on ```Ubuntu 22.04.2 LTS``` and it worked for me.
