### 2024-09-03 dfki-ric/ugv_nav4d#9 @OlgerSiebinga
Automatic build fails if QT 5.15.3 is installed
I tried to run the automatic build on Ubuntu 22.04.4 with QT 5.15.3 installed. The automatic build script caused my OS to freeze in a (seemingly) endless loop.  After some digging in the trace I found that this specific version of QT was the most likely cause: `-- Found unsuitable Qt version "5.15.3" from /usr/bin/qmake` while the script was configuring `gui-vizkit3d`. Manually removing QT5 and rerunning the build script fixed the issue.

Full output until my OS froze: [output.txt](https://github.com/user-attachments/files/16848403/output.txt)

(part of https://github.com/openjournals/joss-reviews/issues/6983)
