### 2024-09-03 dfki-ric/ugv_nav4d#6 @OlgerSiebinga
It is strange indeed. I deleted everything and started over and ran into another issue ( https://github.com/dfki-ric/ugv_nav4d/issues/9). I fixed that one, and the Rock error did not show again. But I really don't understand why it's gone now.
