### 2024-09-03 dfki-ric/ugv_nav4d#10 @OlgerSiebinga
GUI will not run after automatic build
my installation finished without errors, however, the GUI will not run. Calling `ugv_nav4d_bin-qt5 test_data/parking_deck.ply 0.3` returns
```
Warning: Ignoring XDG_SESSION_TYPE=wayland on Gnome. Use QT_QPA_PLATFORM=wayland to run on Wayland anyway.
Calling get/setColor() on SubTrajectoryVisualization is deprecated.
Calling get/setColor() on SubTrajectoryVisualization is deprecated.
terminate called after throwing an instance of 'std::runtime_error'
  what():  gridSize has to be > 0
Aborted (core dumped)
```

(part of https://github.com/openjournals/joss-reviews/issues/6983)
