### 2024-09-04 dfki-ric/ugv_nav4d#9 @haider8645
@planthaber @pierrewillenbrockdfki any idea why this could happen? I also get the same error on rebuild that ```-- Found unsuitable Qt version "5.15.3" from /bin/qmake```
