### 2024-09-04 dfki-ric/ugv_nav4d#9 @pierrewillenbrockdfki
> Found unsuitable Qt version "5.15.3" from /bin/qmake

This is from cmakes FindQt4 macro package, which then correctly reports that qt4 was not found. This could be suppressed by adding QUIET to the FindPackage call.

The original problem with the OS apparently freezing usually is caused by a compiler _using_ more virtual memory than the system can keep present in physical memory, causing the system to swap out lesser used virtual memory, like the users graphical session. @haider8645 does this use a parallel build(`-j`)? Not building parallel can relieve memory pressure simply by not having many compilers trying the use memory.

In the olden days of spinning rust, you'd literally hear the disk thrashing to move the memory pages in and out of swap, but nowadays you only get a lit access light, if anything. You'd then be either waiting and hoping that some work will finish to free up memory, or reach for SysRq+R,E,I,S,U,B.

Also want to mention that this rarely points to a hardware issue where the system overheats or has some bad memory cells. In that case, it would probably not react to SysRq anymore, but there are also reasons why SysRq would not work even though the Linux kernel is still alive(For example, SysRq being disabled).
