### 2024-09-04 dfki-ric/ugv_nav4d#9 @OlgerSiebinga
Do I understand you correctly when I say that with this virtual memory issue, the build would continue in the background? Because in that case, I doubted whether this was the problem.

My OS was frozen the first time for something like 20 minutes. It was stuck at the same step it reached in the output.txt file. However, I obtained this file by redirecting the output of the terminal to a file, and let my machine run for over 2 hours (while it was frozen) in the hope that it was doing something in the background. But it did not get any further this time. I would expect that in the second situation (with the output going straight to a file), it would not be affected by the system taking memory away from the graphical session.

So I'm not completely sure, but I suspect something else is happening.
