### 2024-09-04 dfki-ric/ugv_nav4d#9 @pierrewillenbrockdfki
Your output.txt looks normal to me; the compiler is warning about a few different functions that use a deprecated QFlags constructor, and then output just stops. I can already tell that multiple compilers are in fact running at the same time, since the warning messages do not follow the begin of compilation(`Building CXX object ...`), but instead appear later. I think there are about 8 compilers running in parallel. At around 4GB memory use per compiler, that would mean 32GB physical memory required. Usually, compilers use way less, though.

Since userspace cannot ever truly crash or freeze a system(that would be a kernel or hardware bug), something else must be going on where the system appears frozen but is not(meaning some amount of work is still getting done).

The situation with too little physical memory is one of those situations where the system might appear frozen(The OutOfMemory killer only acts when the swap space is full, too). Another one could be when thousands of processes get created in a short amount of time(see forkbomb).

A single process that does not use a huge amount of memory and does not spawn a huge amount of threads cannot freeze a system, just create a lot of system load, but generally leaves enough time for interactive processes to do their work(processes that have been idle preferentially get time slices assigned when they become runnable).

C++ compilers are known to allocate and use huge amounts of memory(multiple GBs). I have seen systems appear frozen due to that, only that i could see it coming when GUI started to stutter and then freeze, with a system health tool showing a sharp increase in swap traffic, and hear the harddisk thrashing during that time(A few years back, when harddisks were still common).

At that point, the compilation slows by a few orders of magnitude. One can try to wait it out, but that could take days. In that situation, very little real work is done, the kernel is busy swaping pages in and out, while processes wait for their memory to be swapped back in, only to be swapped out when the next process demands its memory.

Since you are using ubuntu, you can try to limit the total amount of virtual memory available to the compilation, but i'd expect the compiler processes to be killed before they are done for using too much memory.

It looks like you can limit the amount of memory available to some program/process group by running it with systemd-run like this:
```
systemd-run --scope -p MemoryMax=100K --user bash
```
The above should not start a shell but instead come up with an error "killed"; if it does instead create a new shell, try without --user, but that requires temporarily elevating privileges(systemd-run on some systems silently fails to set the memory limit). For testing this on a compilation attempt, i guess `MemoryMax=10G` or something in that range would be required. But again, i would expect compilation to abort with this.

I think the real fix is to stop compiling in parallel, which @haider8645 probably knows better how to accomplish.
