### 2024-09-05 dfki-ric/ugv_nav4d#9 @planthaber
Or remove the "-j" here ans try again:
https://github.com/dfki-ric/ugv_nav4d/blob/main/source_dependencies/build.bash#L25
