### 2024-09-05 dfki-ric/ugv_nav4d#9 @OlgerSiebinga
Thanks for the suggestions. To clarify, manually removing Qt5 from my system already solved the issue. Without Qt 5.15.3 installed, the build runs without any problems.
