### 2024-09-05 dfki-ric/ugv_nav4d#10 @haider8645
I tried a completely fresh install and the GUI worked. In your case, I guess the issue is with the parsing of the gridSize. Could you try ```ugv_nav4d_bin-qt5 test_data/parking_deck.ply``` and see if the GUI loads.
