### 2024-09-05 dfki-ric/ugv_nav4d#9 @haider8645
The automatic build works without the parallel build as suggested by @pierrewillenbrockdfki. Thanks!
