### 2024-09-10 dfki-ric/ugv_nav4d#10 @OlgerSiebinga
Yes, then it loads but with a whole bunch of errors:

```
  ugv_nav4d_bin-qt5 test_data/parking_deck.ply
Warning: Ignoring XDG_SESSION_TYPE=wayland on Gnome. Use QT_QPA_PLATFORM=wayland to run on Wayland anyway.
Calling get/setColor() on SubTrajectoryVisualization is deprecated.
Calling get/setColor() on SubTrajectoryVisualization is deprecated.
[20240910-09:54:21:093] [ERROR] - ugv_nav4d::Point 0.978741        0  12.2618 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:093] [ERROR] - ugv_nav4d::Point 1.85393       0 12.2543 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:093] [ERROR] - ugv_nav4d::Point 1.75729       0 8.25864 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:093] [ERROR] - ugv_nav4d::Point 2.67342       0 8.25357 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:094] [ERROR] - ugv_nav4d::Point 1.62613       0 4.26153 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:094] [ERROR] - ugv_nav4d::Point 1.48058       0 4.25463 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:112] [ERROR] - ugv_nav4d::Point       0 30.2746 12.3272 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:112] [ERROR] - ugv_nav4d::Point       0 30.2727 13.9321 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:112] [ERROR] - ugv_nav4d::Point       0 30.2708 15.9005 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:112] [ERROR] - ugv_nav4d::Point       0 30.2733 15.5897 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:112] [ERROR] - ugv_nav4d::Point       0  30.276 7.25286 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:112] [ERROR] - ugv_nav4d::Point       0 30.2772 14.9678 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:112] [ERROR] - ugv_nav4d::Point       0 30.2715 10.7747 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:112] [ERROR] - ugv_nav4d::Point       0 30.2687   14.23 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:112] [ERROR] - ugv_nav4d::Point       0 30.2685 6.91518 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:112] [ERROR] - ugv_nav4d::Point       0  30.272 7.61544 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:112] [ERROR] - ugv_nav4d::Point       0 30.2695 12.0553 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:112] [ERROR] - ugv_nav4d::Point       0 30.2738 6.39793 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:112] [ERROR] - ugv_nav4d::Point       0 30.2698 12.6342 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:114] [ERROR] - ugv_nav4d::Point       0 30.2783 7.61382 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:116] [ERROR] - ugv_nav4d::Point       0 30.2774 7.28304 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:116] [ERROR] - ugv_nav4d::Point       0 30.2741 15.2059 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:116] [ERROR] - ugv_nav4d::Point       0 30.2698 14.1817 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:116] [ERROR] - ugv_nav4d::Point       0 30.2751 14.4689 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:116] [ERROR] - ugv_nav4d::Point       0  30.277 5.15889 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:116] [ERROR] - ugv_nav4d::Point       0  30.276 5.12658 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:116] [ERROR] - ugv_nav4d::Point       0 30.2741 15.5169 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:116] [ERROR] - ugv_nav4d::Point       0 30.2691 10.8507 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:116] [ERROR] - ugv_nav4d::Point       0 30.2701 6.92212 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:116] [ERROR] - ugv_nav4d::Point       0 30.2691 2.50401 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:116] [ERROR] - ugv_nav4d::Point        0  30.2716 0.506143 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:116] [ERROR] - ugv_nav4d::Point       0 30.2741 5.54068 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:116] [ERROR] - ugv_nav4d::Point       0 30.2683 7.86908 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:116] [ERROR] - ugv_nav4d::Point       0 30.2731  7.3687 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:124] [ERROR] - ugv_nav4d::Point   1.463       0 4.25841 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:124] [ERROR] - ugv_nav4d::Point  2.3104       0 4.25949 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:125] [ERROR] - ugv_nav4d::Point 2.01916       0  8.2603 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[20240910-09:54:21:125] [ERROR] - ugv_nav4d::Point 2.17689       0 8.25696 is outside of the grid! Can't add to grid. (/home/localadmin/cpp_projects/ugv_nav4d/build/install/include/maps/grid/MLSMap.hpp:222 - void maps::grid::MLSMap<SurfaceType>::mergePointCloud(const PointCloud&, const Transform3d&, double) [with maps::grid::MLSConfig::update_model SurfaceType = maps::grid::MLSConfig::SLOPE; maps::grid::PointCloud = pcl::PointCloud<pcl::PointXYZ>; base::Transform3d = Eigen::Transform<double, 3, 2, 2>])
[…]
