### 2024-09-10 3650688
Update paper and enable add pimtivies in the debug gui
