+++ paper/paper.md
@@ -23,150 +23,49 @@ bibliography: paper.bib
 ---
 
 # Summary
-Navigation of autonomous mobile robots is a complex problem which comprises of multiple software components, each performing a specific task, which when combined form the navigation software architecture. Ugv_nav4d can be used to navigate complex indoor and outdoor environments. It provides a traversable, motion constraint compliant, and collision free trajectory. It is highly customizable and can be used for any type of terrestrial robot.
 
-To find a traversable trajectory, ugv_nav4d internally maintains a 3D Traversability Map (TraversabilityMap3d) of the environment. The TraversabilityMap3d [@slammaps] is generated based on a multi-layered surface map (MLS [@mlsmaps], [@slammaps]), which enables ugv_nav4d to plan trajectories in multi-surface environments. The trajectory is composed by careful selection and piecing together of motion primitives from a pool of primitives. The primitives are generated based on the specific mechnical features of the robot. As a result, the planned trajectory always fits well with the motion capabilites of the robot. 
-
-In short, ugv_nav4d
-
-- can plan continuous, smooth, and feasible paths in complex indoor and outdoor environments.
-- can plan in multi-surface environments, e.g. natural cave system, multi-storey buildings.
-- provides a complete trajectory for the planned path which is ready for execution.
-- provides a graphical user interface (GUI) for introspection and debugging.
+The ugv_nav4d is a global path planner designed for terrestrial robots to navigate complex indoor and outdoor environments. To achieve this, ugv_nav4d uses a traversability map (TraversabilityMap3d) based on a multi-layered surface map (MLSMap) [@mlsmaps], [@slammaps], enabling planning in multi-surface environments. The path is built by selecting motion primitives tailored to the robot's mechanical features, ensuring compatibility with its motion capabilities.
 
 ![Planned trajectory in a multi-storey environment.](figures/parking_deck.png){width="300pt"}
 
 # Statement of need
+Accurate ground surface representation is crucial for ground-based robots in complex terrains. The ROS Navigation Stack (nav2) [@ros2], which uses voxel maps for 3D navigation, often loses detail and accuracy, especially in multi-storey environments, due to its discrete voxelization and separate costmaps for each floor.
 
-The ROS navigation stack [@ros2] is undoubtedly the most used open-source robotics framework. This claim can be backed by its widespread usage in research and industry. Navigation2 (nav2) [@macenski2020nav2] is the newest navigation system which is based on the legacy of ROS Navigation. In this paper, we propose a new, fully open-source robot navigation library called ugv_nav4d. Ugv_nav4d presents significant improvements over nav2, particularly in environmental representation and navigation in complex, uneven terrains and multistorey environments. By utilizing Multi-Layered Surface Maps [@mlsmaps] and a 3D Traversability Map [@slammaps], ugv_nav4d offers a more detailed and accurate navigation solution for ground vehicles. These capabilities are essential for operations in challenging terrains and address critical needs within the robotics community. The ongoing development of a ROS2 wrapper for ugv_nav4d aims to facilitate its adoption and integration within the ROS2 ecosystem, providing substantial benefits to the community and enhancing the capabilities of ground vehicle navigation.
+We propose ugv_nav4d, a path planner that enhances environmental representation with Multi-Layered Surface Maps (MLS) [@mlsmaps] and a 3D Traversability Map [@slammaps]. Ugv_nav4d avoids the "stepping" effect of voxel maps by using a continuous grid and detailed vertical information, providing smoother and more accurate terrain modeling.
 
-### Environment Representation and Mapping Techniques
[…]
