### 2024-09-10 dfki-ric/ugv_nav4d#10 @haider8645
That just means that some points of the PLY file could be added to the MLS map because of being just on the borders of the grid. But yeah for a user that would looks like something is broken.
