### 2024-09-10 33be466
Add acknowledgement for contributors and fix figure
