### 2024-09-10 #6983 @haider8645
@OlgerSiebinga please find the updated paper. Took all your recommendations into consideration. The word count is now down from over 2000 to under 1000. The writing style is updated. Added acknowledgement for the contributors and combined figures etc.
