### 2024-09-10 dfki-ric/ugv_nav4d#10 @chhtz
I think that this should just be a warning (it only concerns a few points at the border of the map and should not influence the navigation). But I'm working on a fix for that.
