### 2024-09-10 #6983 @logological
@haider8645 Thanks for letting us know about the revisions.  @soraxas I don't believe we've heard from you since 17 July – could you please confirm whether the issues you've raised have been addressed to your satisfaction?  If so, perhaps you could also now move on to the other items in your checklist.  Please remember to work from the latest draft of the paper, as the authors have already made some substantial revisions based on the other reviewer's comments.
