### 2024-09-17 dfki-ric/ugv_nav4d#7 @soraxas
Looks good
