### 2024-09-17 dfki-ric/ugv_nav4d#3 @soraxas
Nice, and it seems you've removed the old blob data by cleaning the repo 👍
