### 2024-09-17 dfki-ric/ugv_nav4d#6 @soraxas
My main system is Arch, and I have numerous issues with getting the correct dependencies as listed in [install_os_dependencies.bash](https://github.com/dfki-ric/ugv_nav4d/blob/main/source_dependencies/install_os_dependencies.bash). (hence the delayed review). I wasn't quite able to get the correct listed of dependencies as the cmake expects certain versions of packages and has version incompatibilities across packages (and the `apt-get ...` commands doesn't quite specify the needed versions).

In the end, I've given up and ended up using a fresh copy of container with Ubuntu 22 that can run the `apt-get` commands. I can however confirm that a fresh ubuntu 22 did not encounter the issue noted by @OlgerSiebinga, and installs without issues. (so perhaps the noted issue is due to conflicts on existing systems, like qt as noted in another issues).
