### 2024-09-17 dfki-ric/ugv_nav4d#2 @soraxas
Looks good to me.
