### 2024-09-17 dfki-ric/ugv_nav4d#8 @soraxas
Some C++ examples:

https://github.com/tbrown122387/pf/tree/master/test

https://github.com/kdm9/libqcpp/tree/master/src/tests

Both of these are JOSS C++ libraries
