### 2024-09-17 #6983 @haider8645
> Hi @haider8645, With your revision, the paper really improved in my opinion. Some other updates from my side:
>
> * I’ve also checked the authorship box now.
> * One open question was about the ROS2 integration. Do you plan to include that in the published version or is it future work?
> * I got the code running but still have to verify if the QT5.15 bug is now solved.
>
> In the meantime, I’ve moved on to the documentation (i.e., the readme file and auto-generated doc) and I have some remarks. Please find them below in order of the checklist/appearance in the docs:
>
> * The statement of need is missing from the docs
> * The figures in the doxygen docs are not working (they are in ugv_nav4d/doc/figures/ while the file points to ugv_nav4d/build/doc/figures/)
> * In the readme, in the Planning section, there is an open TODO.
> * The example usage is based on the (debug) gui with a provided map.  Although great for quickly getting a feel for ugv_nav4d can do, I’m still unsure how/where to start if I wanted to use your software on your robot. I think such a real-world example could greatly help new users to get started. Suppose I have a robot and a setup in the lab. What would I need to do to get started. What information would go in and what would come out. Both in terms of GUI usage and API calls on the real robot. The review checklist ask for an example “ideally to solve real-world analysis problems”. So I guess technically you don’t necessarily need it, but I would highly recommend such an example.
>
> I plan to dive a bit deeper into the functionality next week.

![image](https://github.com/user-attachments/assets/d053d438-7c6f-4a75-90ab-8412dcc625e0)

ROS2 integration is looking good. [Working on it ](https://github.com/dfki-ric/ugv_nav4d/issues/1) parallel to the review here.

I will work on the other points you mentioned about the missing statement of need and the real-world-example. I guess this is the ROS2 wrapper would be a good starting point. The open TODO is in the readme is an old artifact, I can remove it.

About the real world usage: There is a known challenge for mapping  in dynamic environments (which is out of scope of the planner and more of a mapping problem), that the map (MLS) is generated based on a pointcloud and the pointcloud should not have any dynamic objects because then they will appear in the map as ghost artifacts. So for the planner to work in dynamic environments, one would need to pre-process the pointcloud and remove dynamic points in so the map is clean of dynamic objects and suitable for the planning.
