### 2024-09-17 dfki-ric/ugv_nav4d#8 @planthaber
The tests are located here: https://github.com/dfki-ric/ugv_nav4d/tree/main/src/test

@haider8645: Add the TESTS_ENABLED=ON cmake option, ply download and running the tests in the github action.
