### 2024-09-17 dfki-ric/ugv_nav4d#10 @haider8645
The errors about out-of-bound points should not appear anymore. @chhtz thanks!
