### 2024-09-17 dfki-ric/ugv_nav4d#8 @soraxas
Thanks @planthaber but that's what I've already linked in the OP.

As mentioned, some form of unit test would be better (there's only one file in the folder, contains only integration test, i.e., end-to-end test).

E.g. From a quick glance, (non-exhaustive list) your repo has things like Dijkstra algorithm, EnvironmentXYZTheta, compute motion, etc. These are individual item can be use some form tests.
