### 2024-09-17 058c57b
Add names for the visualizations in debug gui and disable 2d trajectory
