### 2024-09-17 dfki-ric/ugv_nav4d#8 @haider8645
> ply download

uploaded the ply files to zenodo. They just need to be added to your organization by the curators and then can remove the test files from repo and update docu and action
