### 2024-09-17 f145930
PCL 1.14 stops adding the version number to the pkg-config files
