### 2024-09-17 dfki-ric/ugv_nav4d#6 @soraxas
Since I've verified that the install script works on a fresh Ubuntu 22, and it seems it works for @OlgerSiebinga after removing his previous QT installation, let's close this .
