### 2024-09-19 4ff9bc0
Disable grid visualization by default and update label for max time
