### 2024-09-19 1258fe0
Update Readme to get test map from zenodo record
