### 2024-09-19 dfki-ric/ugv_nav4d#10 @chhtz
> terminate called after throwing an instance of 'std::runtime_error'
>   what():  gridSize has to be > 0

The cause for that appears to be that `atof` takes the current locale for parsing numbers. It should work if you pass `0,3` instead of `0.3`, I assume.
