### 2024-09-20 dfki-ric/ugv_nav4d#10 @OlgerSiebinga
Using 0,3 works for me indeed!

It made me curious though, what does this parameter do? (and are there other parameters I can use when calling the gui from the terminal?) I cannot find anything about it in the documentation other then the example. But I might me missing something.
