### 2024-09-20 #6983 @OlgerSiebinga
>
> ![image](https://private-user-images.githubusercontent.com/23505408/368091293-d053d438-7c6f-4a75-90ab-8412dcc625e0.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjY4MjY0MjAsIm5iZiI6MTcyNjgyNjEyMCwicGF0aCI6Ii8yMzUwNTQwOC8zNjgwOTEyOTMtZDA1M2Q0MzgtN2M2Zi00YTc1LTkwYWItODQxMmRjYzYyNWUwLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA5MjAlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwOTIwVDA5NTUyMFomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPWIyZWE2NTM5YTg0NzAwNjMwMGVhNjk2N2JhY2NiODE4ZWNiZTczY2I0YWUyMzg3NTJkZjU2Y2FhNWIwNDZhOTQmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.erfcKlPOM9bsq0Zf-eN3DWSYEWgyl9HEyjvA3htLM9o)
>
> ROS2 integration is looking good. [Working on it ](https://github.com/dfki-ric/ugv_nav4d/issues/1) parallel to the review here.
>
> I will work on the other points you mentioned about the missing statement of need and the real-world-example. I guess this where the ROS2 wrapper would be a good starting point. The open TODO is in the readme is an old artifact, I can remove it.
>
> Just something extra (about the real world usage): There is a known challenge for mapping in dynamic environments (which is out of scope of the planner and more of a mapping problem), that the map (MLS) is generated based on a pointcloud and the pointcloud should not have any dynamic objects because then they will appear in the map as ghost artifacts. So for the planner to work in dynamic environments, one would need to pre-process the pointcloud and remove any dynamic points from moving objects so that the map is clean and suitable for the planning. The detection of dynamic points in point clouds is ongoing research: https://github.com/PRBonn/4DMOS

The ROS integration look really nice! I think the extra information on real-world usage would be very valuable in the documentation. When combined with a rough outline on which steps to take, I think that would be enough information for potential users to get started.
