### 2024-09-20 dfki-ric/ugv_nav4d#10 @chhtz
The `0.3` is the grid-size. With smaller values, planning can be finer (unless it is too small to create a usable map, i.e. smaller than the resolution of the point cloud), but generally also may take longer. E.g., `0.1` still works for me. Values `0.5` and `1.0` seem to be prohibited and I did not succeed generating paths with values above that (maybe some other parameters need to be changed to make that work).

Most parameters are hardcoded here:
https://github.com/dfki-ric/ugv_nav4d/blob/main/src/gui/PlannerGui.cpp#L294

As it is, the GUI is more a demonstrator. One could parse additional parameters, or read them from a config file.
