### 2024-10-04 #6983 @logological
Just wanted to check in with everyone regarding the status of the review.  I see there are still some open issues in the project repository; are these currently blocked by external factors or are we waiting on the submitters/reviewers for further action?
