### 2024-10-07 ab34f4d
Do not replan during planning phase in gui
