### 2024-10-09 #6983 @OlgerSiebinga
Hi @logological,

My only open checkbox is the example usage. Once that is updated in the documentation, the paper can be accepted in my opinion. Although I'm still a fan of including the ROS integration in the published version of the software (which would mean that I'll have to review that functionality as well).
