### 2024-10-10 abdfab3
Enable forward primitives by default
