+++ src/gui/PlannerGui.cpp
@@ -302,8 +302,8 @@ void PlannerGui::setupPlanner(int argc, char** argv)
     splineConfig.numEndAngles = 21;
     splineConfig.destinationCircleRadius = 10;
     splineConfig.cellSkipFactor = 3.0;
-    splineConfig.generatePointTurnMotions = true;
-    splineConfig.generateLateralMotions = true;
+    splineConfig.generatePointTurnMotions = false;
+    splineConfig.generateLateralMotions = false;
     splineConfig.generateBackwardMotions = true;
     splineConfig.generateForwardMotions = true;
     splineConfig.splineOrder = 4;
