### 2024-10-11 #6983 @haider8645
Hi  @logological

I am working on the example usage as requested by @OlgerSiebinga. THe ROS2 integration is looking good. A planned path on the traversability map can be seen in rviz2. I am testing the ros2 wrapper and will publish it afterwards in the coming days.

![image](https://github.com/user-attachments/assets/48b3f98e-c72d-47cb-880c-b3f5cf931d1c)
