### 2024-10-11 22446a6
Throw exception when sizes of costs to start and goal do not match
