### 2024-10-11 fd10981
Update GUI parameters
