+++ src/gui/PlannerGui.cpp
@@ -298,12 +298,12 @@ void PlannerGui::setupPlanner(int argc, char** argv)
          res = atof(argv[2]);
     
     splineConfig.gridSize = res;
-    splineConfig.numAngles = 42;
-    splineConfig.numEndAngles = 21;
-    splineConfig.destinationCircleRadius = 10;
-    splineConfig.cellSkipFactor = 3.0;
-    splineConfig.generatePointTurnMotions = false;
-    splineConfig.generateLateralMotions = false;
+    splineConfig.numAngles = 16;
+    splineConfig.numEndAngles = 8;
+    splineConfig.destinationCircleRadius = 6;
+    splineConfig.cellSkipFactor = 0.1;
+    splineConfig.generatePointTurnMotions = true;
+    splineConfig.generateLateralMotions = true;
     splineConfig.generateBackwardMotions = true;
     splineConfig.generateForwardMotions = true;
     splineConfig.splineOrder = 4;
@@ -311,15 +311,15 @@ void PlannerGui::setupPlanner(int argc, char** argv)
     mobilityConfig.translationSpeed = 0.5;
     mobilityConfig.rotationSpeed = 0.5;
     mobilityConfig.minTurningRadius = 1; // increase this to reduce the number of available motion primitives
-    mobilityConfig.searchRadius = 0.0;
-    mobilityConfig.multiplierForward = 1;
-    mobilityConfig.multiplierBackward = 3;
-    mobilityConfig.multiplierLateral = 4;
-    mobilityConfig.multiplierBackwardTurn = 4;
-    mobilityConfig.multiplierLateralCurve = 4;
-    mobilityConfig.multiplierForwardTurn = 2;
-    mobilityConfig.multiplierPointTurn = 3;
+    mobilityConfig.searchRadius = 1.0;
     mobilityConfig.searchProgressSteps = 0.1;
+    mobilityConfig.multiplierForward = 1;
+    mobilityConfig.multiplierForwardTurn = 2;
+    mobilityConfig.multiplierBackward = 2;
+    mobilityConfig.multiplierBackwardTurn = 3;
+    mobilityConfig.multiplierLateral = 4;
+    mobilityConfig.multiplierLateralCurve = 4;
+    mobilityConfig.multiplierPointTurn = 3;
     mobilityConfig.maxMotionCurveLength = 100;
     mobilityConfig.spline_sampling_resolution = 0.05;
     mobilityConfig.remove_goal_offset = false;
