### 2024-10-16 908408c
Add example case to readme
