+++ README.md
@@ -135,7 +135,10 @@ sudo apt install ros-humble-desktop-full
 ### 2. Install Gazebo Fortress
 If you need to install **Gazebo Fortress**, follow the instructions provided on the official page at [Gazebo Installation](https://gazebosim.org/docs/latest/ros_installation/).
 
-### 3. Get the repos 
+### 3. Install SLAM
+If you have a SLAM package which provides a pointcloud map on a topic then you can skip this step. If not then you can use [lidarslam_ros2](https://github.com/rsasaki0109/lidarslam_ros2). Please follow the build and install instructions from the original repository. Set the parameter `robot_frame_id: "husky/base_link"` for the `scanmatcher` node in [lidarslam.yaml](https://github.com/rsasaki0109/lidarslam_ros2/blob/a63b8fa2485e05251505b2bb209598285106bff2/lidarslam/param/lidarslam.yaml#L4)
+
+### 3. Get ugv_nav4d_ros2 and a test environment for robot husky in gazebo
 
 ```
 mkdir -p your_ros2_workspace/src
@@ -146,6 +149,8 @@ You can clone the repo `ros2_humble_gazebo_sim` anywhere in your system. Here we
 ```
 cd ~/your_ros2_workspace
 git clone https://github.com/dfki-ric/ros2_humble_gazebo_sim.git
+cd ros2_humble_gazebo_sim
+bash install_dependencies.bash
 ```
 ### 4. Husky Configuration
 To ensure that Gazebo can find the robot model, you need to export the following environment variable. Replace /path/to/ with the actual **complete** path where you clone the repository `ros2_humble_gazebo_sim`. Add this command to your terminal:
@@ -154,7 +159,7 @@ export IGN_GAZEBO_RESOURCE_PATH=/path/to/your_ros2_workspace/ros2_humble_gazebo_
 ```
 
 ### 5. Building the ROS 2 Workspace
-Before launching the simulation, build your ROS 2 workspace with the release flag:
+Before launching the simulation, build your ROS 2 workspace:
 
 ```
 cd ~/your_ros2_workspace
@@ -168,16 +173,99 @@ source ~/your_ros2_workspace/install/setup.bash
 cd ~/your_ros2_workspace/ros2_humble_gazebo_sim/simulation
 ros2 launch start.launch.py
 ```
+You can use the `Teleop` plugin of Gazebo for sending velocity commands to the robot. Click on the three dots in top-right corner of Gazebo window and search for `Teleop`. Select the plugin and adjust the values as shown in figure.
 
-In a new terminal, source your workspace, ugv_nav4d library, and launch the ugv_nav4d_ros2. Replace the /path/to/ugv_nav4d with the location of the ugv_nav4d library. Add this command to your terminal:
+![GazeboTeleop](doc/figures/gazebo_teleop.png)
+
+Alternative to the `Teleop` plugin, you can use a joystick for moving the robot. For this, set the argument `use_joystick:=true`. Adjust the values in the folder `/config` for your joystick setup and provide the full paths to the arguments `joy_config_file` and `teleop_twist_config_file` as shown below:
+
+```
+ros2 launch start.launch.py use_joystick:=True joy_config_file:=/your_ros2_workspace/ros2_humble_gazebo_sim/simulation/config/joy_config.yaml teleop_twist_config_file:=/your_ros2_workspace/ros2_humble_gazebo_sim/simulation/config/teleop_twist_config.yaml
+```
+
[…]
