### 2024-10-16 dfki-ric/ugv_nav4d#16 @haider8645
Example Case
Hello @OlgerSiebinga,

I added an [example case in the documentation](https://github.com/dfki-ric/ugv_nav4d/tree/example_case?tab=readme-ov-file#ros-2-humble-test-environment-with-gazebo-fortress). It shows how to use ugv_nav4d in ros2 within a test environment. The repositories for the test case are not yet published. The example should give a good starting point for users.

I will let you know after publishing the ros2 wrapper so that you can test it.
