### 2024-10-16 dfki-ric/ugv_nav4d#16 @haider8645
@soraxas FYI
