### 2024-10-16 4084406
Minor adjustments in readme
