### 2024-10-16 6f80bca
Add figure for urban_circuit
