### 2024-10-24 dfki-ric/ugv_nav4d#16 @haider8645
The repositories are published.

https://github.com/dfki-ric/ugv_nav4d_ros2
https://github.com/dfki-ric/ros2_humble_gazebo_sim
