### 2024-10-25 dfki-ric/ugv_nav4d#16 @soraxas
Nice!
