### 2024-10-28 #6983 @haider8645
@logological @OlgerSiebinga Just a heads up that I have added the [example case](https://github.com/dfki-ric/ugv_nav4d/issues/16).
