### 2024-10-31 dfki-ric/ugv_nav4d#16 @OlgerSiebinga
HI @haider8645,

I tried to run the example case in ros on a fresh installation today, but I got stuck at step 5 of the readme instructions: the ros2 build. It seems there is a dependency missing in the instruction because I get the following error message:

```
$ colcon build --cmake-args -DCMAKE_BUILD_TYPE=Release
Starting >>> lidarslam_msgs
Starting >>> ndt_omp_ros2
Starting >>> ugv_nav4d_ros2
Finished <<< ndt_omp_ros2 [0.25s]
--- stderr: ugv_nav4d_ros2
CMake Error at /usr/share/cmake-3.22/Modules/FindPkgConfig.cmake:603 (message):
  A required package was not found
Call Stack (most recent call first):
  /usr/share/cmake-3.22/Modules/FindPkgConfig.cmake:825 (_pkg_check_modules_internal)
  CMakeLists.txt:19 (pkg_check_modules)

---
Failed   <<< ugv_nav4d_ros2 [0.25s, exited with code 1]
Aborted  <<< lidarslam_msgs [0.34s]

Summary: 1 package finished [0.61s]
  1 package failed: ugv_nav4d_ros2
  1 package aborted: lidarslam_msgs
  1 package had stderr output: ugv_nav4d_ros2
  3 packages not processed
```
