### 2024-10-31 dfki-ric/ugv_nav4d#16 @OlgerSiebinga
turns out I had to source the `env` of the original ugv_nav4d installation to make it visible to Ros. This is mentioned in the wrapper readme but not in the example instructions. Might be nice to add it there.
