### 2024-11-01 dfki-ric/ugv_nav4d#16 @OlgerSiebinga
I ran into another minor thing, I had to manually install `ros-humble-ros-ign` to get the launch script to run. After that, I managed to  get the launch script to run, but returns the following error:

```
[Err] [Server.cc:139] Error Code 13: [/sdf/world[@name="gz"]/include[0]/uri:/home/localadmin/ros2_ws/ros2_humble_gazebo_sim/simulation/models/cave_circuit.sdf:L16]: Msg: Unable to find uri[model://COSTAR_HUSKY_SENSOR_CONFIG_1]
```
Looks like I'm missing a config file. Or is this something I can fix on my side?
