### 2024-11-01 dfki-ric/ugv_nav4d#16 @haider8645
Do you get the same error even after sourcing the following:

```
export IGN_GAZEBO_RESOURCE_PATH=/home/localadmin/ros2_ws/ros2_humble_gazebo_sim/resource:$IGN_GAZEBO_RESOURCE_PATH
```
