### 2024-11-01 dfki-ric/ugv_nav4d#16 @OlgerSiebinga
thanks, I forgot that one!
The list of `source`s and `export`s to run the launch command is quite long. It might be nice to provide an extensive list because now I had to trace it back from between the build commands in the example (I ran all the builds yesterday and just wanted to launch the example today). Anyway, just a suggestion.
