### 2024-11-01 dfki-ric/ugv_nav4d#16 @haider8645
Yes, I agree with you. It would be a challenge for anyone to go through all the commands and text. Hmm, I will find a way to make it user friendly. Thanks for the heads up!
