### 2024-11-01 #6983 @OlgerSiebinga
I verified that the example usage and the ROS2 wrapper are working. I think this extra effort really payed off to make the software easy to get started and play around with for new users.

I've checked the final checkbox on my review list. So I think the paper is ready to be accepted for publication
