### 2024-11-01 dfki-ric/ugv_nav4d#8 @haider8645
@soraxas I am working on the unit tests.
