### 2024-11-01 dfki-ric/ugv_nav4d#10 @haider8645
@OlgerSiebinga can this be closed?
