### 2024-11-01 dfki-ric/ugv_nav4d#10 @OlgerSiebinga
yes, I think it works now. So if the example now show 0,3 instead of 0.3 (so it runs when you copy it) I think the issue can be closed
