### 2024-11-01 #6983 @logological
Thanks, @OlgerSiebinga! @soraxas, there are a couple items left unchecked in your review.  Could you have a look at these and let me and @haider8645 know if any further updates to the paper or the project are required?
