### 2024-11-04 dfki-ric/ugv_nav4d#10 @chhtz
Regarding `0.3` vs `0,3`, either we should replace `atof` with something which ignores the locale or add a note that the locale may influence the behavior (I'd prefer the first, honestly).
