### 2024-11-05 adcc0c2
Add unit tests for environmentxyztheta
