### 2024-11-13 2574ca4
Update handling of planning in gui
