+++ src/test/test_EnvironmentXYZTheta.cpp
@@ -19,112 +19,251 @@
 #include <boost/archive/binary_iarchive.hpp>
 
 using namespace ugv_nav4d;
+using namespace maps::grid;
 
-std::string mlsBinFile;
+std::vector<Eigen::Vector3d> startPositions;
+std::vector<Eigen::Vector3d> goalPositions;
 
-// Define EnvironmentXYZThetaTest fixture for Boost Test
 struct EnvironmentXYZThetaTest {
-    EnvironmentXYZThetaTest();
-    ~EnvironmentXYZThetaTest();
+    EnvironmentXYZThetaTest(){}
+    ~EnvironmentXYZThetaTest(){}
 
-    void loadMlsMap();
-
-    typedef EnvironmentXYZTheta::MLGrid MLSBase;
 
     EnvironmentXYZTheta* environment;
-    maps::grid::MLSMapSloped mlsMap;
     Mobility mobility;
     sbpl_spline_primitives::SplinePrimitivesConfig splinePrimitiveConfig;
     traversability_generator3d::TraversabilityConfig traversabilityConfig;
     bool map_loaded = false;
 };
 
-EnvironmentXYZThetaTest::EnvironmentXYZThetaTest() {
-    std::cout << "Called SetUp()" << std::endl;
+BOOST_FIXTURE_TEST_CASE(travmap_resolution_not_equal_to_mls_resolution, EnvironmentXYZThetaTest) {
+    Vector2d res(0.3, 0.3);
+    Vector2ui numCells(10, 10);
 
-    splinePrimitiveConfig.gridSize = 0.3;
-    splinePrimitiveConfig.numAngles = 16;
-    splinePrimitiveConfig.numEndAngles = 8;
-    splinePrimitiveConfig.destinationCircleRadius = 6;
-    splinePrimitiveConfig.cellSkipFactor = 0.1;
-    splinePrimitiveConfig.splineOrder = 4.0;
+    MLSConfig mls_config;
+    mls_config.gapSize = 0.1;
+    mls_config.updateModel = MLSConfig::SLOPE;
+    MLSMapSloped mls_o = MLSMapSloped(numCells, res, mls_config);
 
-    mobility.translationSpeed = 0.5;
-    mobility.rotationSpeed = 0.5;
-    mobility.minTurningRadius = 1;
-    mobility.spline_sampling_resolution = 0.05;
-    mobility.remove_goal_offset = true;
-    mobility.multiplierForward = 1;
-    mobility.multiplierBackward = 3;
-    mobility.multiplierPointTurn = 3;
-    mobility.multiplierLateral = 4;
-    mobility.multiplierForwardTurn = 2;
-    mobility.multiplierBackwardTurn = 4;
-    mobility.multiplierLateralCurve = 4;
-    mobility.searchRadius = 1.0;
-    mobility.searchProgressSteps = 0.1;
-    mobility.maxMotionCurveLength = 100;
+    /** Translate the local frame (offset) **/
+    mls_o.getLocalFrame().translation() << 0.5*mls_o.getSize(), 0;
 
-    traversabilityConfig.maxStepHeight = 0.25;
-    traversabilityConfig.maxSlope = 0.45;
-    traversabilityConfig.inclineLimittingMinSlope = 0.2;
-    traversabilityConfig.inclineLimittingLimit = 0.1;
-    traversabilityConfig.costFunctionDist = 0.0;
-    traversabilityConfig.minTraversablePercentage = 0.4;
-    traversabilityConfig.robotHeight = 1.2;
-    traversabilityConfig.robotSizeX = 1.35;
-    traversabilityConfig.robotSizeY = 0.85;
-    traversabilityConfig.distToGround = 0.0;
[…]
