+++ src/test/test_Planner.cpp
@@ -4,41 +4,33 @@
 #include <fstream>
 #include <cstdlib>
 
-#include "ugv_nav4d/DiscreteTheta.hpp"
 #include "ugv_nav4d/Planner.hpp"
 #include <traversability_generator3d/TraversabilityConfig.hpp>
 #include <maps/grid/MLSMap.hpp>
 
-#include <pcl/io/ply_io.h>
-#include <pcl/common/common.h>
-#include <pcl/common/transforms.h>
-
 using namespace ugv_nav4d;
-
-std::string filePath;
 
 class PlannerTest {
 public:
     PlannerTest();
     ~PlannerTest();
-    void loadMlsMap(const std::string& path);
     std::string getResult(const Planner::PLANNING_RESULT& result);
 
     Planner* planner;
-    maps::grid::MLSMapSloped mlsMap;
+    maps::grid::MLSMapSloped* mlsMap;
     PlannerConfig plannerConfig;
     Mobility mobility;
     sbpl_spline_primitives::SplinePrimitivesConfig splinePrimitiveConfig;
     traversability_generator3d::TraversabilityConfig traversabilityConfig;
-    bool map_loaded = false;
 };
 
 PlannerTest::PlannerTest() {
+
     splinePrimitiveConfig.gridSize = 0.3;
-    splinePrimitiveConfig.numAngles = 42;
-    splinePrimitiveConfig.numEndAngles = 21;
-    splinePrimitiveConfig.destinationCircleRadius = 10;
-    splinePrimitiveConfig.cellSkipFactor = 3;
+    splinePrimitiveConfig.numAngles = 16;
+    splinePrimitiveConfig.numEndAngles = 8;
+    splinePrimitiveConfig.destinationCircleRadius = 6;
+    splinePrimitiveConfig.cellSkipFactor = 0.1;
     splinePrimitiveConfig.splineOrder = 4.0;
 
     mobility.translationSpeed = 0.5;
@@ -47,25 +39,25 @@ PlannerTest::PlannerTest() {
     mobility.spline_sampling_resolution = 0.05;
     mobility.remove_goal_offset = true;
     mobility.multiplierForward = 1;
-    mobility.multiplierBackward = 3;
+    mobility.multiplierBackward = 2;
     mobility.multiplierPointTurn = 3;
     mobility.multiplierLateral = 4;
     mobility.multiplierForwardTurn = 2;
-    mobility.multiplierBackwardTurn = 4;
+    mobility.multiplierBackwardTurn = 3;
     mobility.multiplierLateralCurve = 4;
-    mobility.searchRadius = 0.0;
+    mobility.searchRadius = 1.0;
     mobility.searchProgressSteps = 0.1;
     mobility.maxMotionCurveLength = 100;
 
-    traversabilityConfig.maxStepHeight = 0.25;
+    traversabilityConfig.maxStepHeight = 0.2;
     traversabilityConfig.maxSlope = 0.45;
     traversabilityConfig.inclineLimittingMinSlope = 0.2;
     traversabilityConfig.inclineLimittingLimit = 0.1;
     traversabilityConfig.costFunctionDist = 0.0;
     traversabilityConfig.minTraversablePercentage = 0.4;
-    traversabilityConfig.robotHeight = 1.2;
-    traversabilityConfig.robotSizeX = 1.35;
-    traversabilityConfig.robotSizeY = 0.85;
+    traversabilityConfig.robotHeight = 0.5;
+    traversabilityConfig.robotSizeX = 0.5;
+    traversabilityConfig.robotSizeY = 0.5;
     traversabilityConfig.distToGround = 0.0;
     traversabilityConfig.slopeMetricScale = 1.0;
[…]

[…]
