### 2024-11-13 dc43700
Use set locale to C for grid resolution parameter from cli for gui
