### 2024-11-13 dacc121
Add a hint to source ugv_nav4d before building ros2 workspace
