### 2024-11-13 dfki-ric/ugv_nav4d#16 @haider8645
I have updated the readme with the hint to source ugv_nav4d before building the ros2 workspace. I will close this for now so that a starting example case is available for users and later update the readme to improve overall usability
