### 2024-11-13 dfki-ric/ugv_nav4d#8 @soraxas
Will review, thanks
