### 2024-11-14 24d00a0
Add unit tests for computed motions
