+++ src/test/test_PreComputedMotions.cpp
@@ -33,7 +33,118 @@ BOOST_AUTO_TEST_CASE(generate_motions) {
     mobilityConfig.minTurningRadius = 1.0;
     
     PreComputedMotions* motions = new PreComputedMotions(splinePrimitiveConfig,mobilityConfig);
-    motions->computeMotions(splinePrimitiveConfig.gridSize, 0.3);
+    motions->computeMotions(splinePrimitiveConfig.gridSize, splinePrimitiveConfig.gridSize);
+    delete motions;
+}
+
+BOOST_AUTO_TEST_CASE(check_motions) {
+
+    sbpl_spline_primitives::SplinePrimitivesConfig splinePrimitiveConfig;
+    Mobility mobilityConfig;
+    mobilityConfig.minTurningRadius = 1.0;
+    
+    PreComputedMotions* motions = new PreComputedMotions(splinePrimitiveConfig,mobilityConfig);
+    motions->computeMotions(splinePrimitiveConfig.gridSize, splinePrimitiveConfig.gridSize);
+
+    //We generate primitives for 16 start angles.
+    for (int i = 1; i <= 16; ++i) {
+        DiscreteTheta theta(i,16);
+        std::vector<Motion> motions_start = motions->getMotionForStartTheta(theta);
+        BOOST_REQUIRE(motions_start.size() > 0);
+        for (const auto& motion : motions_start) {
+            BOOST_REQUIRE(motion.baseCost > 0);
+            BOOST_REQUIRE(motion.costMultiplier > 0);
+            BOOST_REQUIRE(motion.costScaleFactor > 0);
+
+            switch (motion.type){
+                case Motion::MOV_FORWARD:
+                    BOOST_CHECK_GT(motion.fullSplineSamples.size(), 0);
+                    BOOST_CHECK_GT(motion.intermediateStepsTravMap.size(), 0);
+                    BOOST_CHECK_GT(motion.intermediateStepsObstMap.size(), 0);
+                    break;
+                case Motion::MOV_BACKWARD:
+                    BOOST_CHECK_GT(motion.fullSplineSamples.size(), 0);
+                    BOOST_CHECK_GT(motion.intermediateStepsTravMap.size(), 0);
+                    BOOST_CHECK_GT(motion.intermediateStepsObstMap.size(), 0);
+
+                    break;
+                case Motion::MOV_LATERAL:
+                    BOOST_CHECK_GT(motion.fullSplineSamples.size(), 0);
+                    BOOST_CHECK_GT(motion.intermediateStepsTravMap.size(), 0);
+                    BOOST_CHECK_GT(motion.intermediateStepsObstMap.size(), 0);
+                    BOOST_CHECK_EQUAL(motion.startTheta.getRadian(), 
+                                      motion.endTheta.getRadian());
+                    break;
+                case Motion:: MOV_POINTTURN:
+                    BOOST_CHECK_EQUAL(motion.fullSplineSamples.size(), 0);
+                    BOOST_CHECK_NE(motion.startTheta.getRadian(), motion.endTheta.getRadian());
+                    break;
+                default:
+                    throw std::runtime_error("test_PreComputedMotions: Invalid motion type detectd!");
+                    break;
+
+            }
+
+            //Point-turns do not have any splines
+            if (motion.type != Motion::MOV_POINTTURN){
+                for (const auto& cellWithPoses : motion.fullSplineSamples){
+                    BOOST_CHECK_GT(cellWithPoses.poses.size(), 0);
+                }
+
+                int splineSize = motion.fullSplineSamples.size();
[…]
