### 2024-11-14 bbee767
Test costs calculation for only one motion
