### 2024-11-14 b6c373d
Check if final pose of full and sampled splines
