### 2024-11-15 cfa92ab
Check validity of underlying values in generated trajectories
