### 2024-11-15 7e4fb90
Remove unnecessary ugv_nav4d-qt5 variant (the library does not depend on qt directly)
