+++ src/gui/CMakeLists.txt
@@ -46,7 +46,7 @@ if(ROCK_QT_VERSION_5)
             PlannerGui.h
         MOC5
             PlannerGui.h
-        DEPS ugv_nav4d-qt5 Qt5::Core
+        DEPS ugv_nav4d Qt5::Core
         DEPS_PKGCONFIG vizkit3d-qt5
                        vizkit3d-viz-qt5
                        maps-viz-qt5
@@ -61,13 +61,13 @@ if(ROCK_QT_VERSION_5)
     rock_executable(ugv_nav4d_bin-qt5
         SOURCES
             Main.cpp
-        DEPS ugv_nav4d-qt5 ugv_nav4d_gui-qt5
+        DEPS ugv_nav4d ugv_nav4d_gui-qt5
     )
 
     rock_executable(ugv_nav4d_replay-qt5
         SOURCES
             ReplayDump.cpp
-        DEPS ugv_nav4d-qt5 ugv_nav4d_gui-qt5
+        DEPS ugv_nav4d ugv_nav4d_gui-qt5
         DEPS_PKGCONFIG vizkit3d-qt5  maps-viz-qt5 base-viz-qt5
         sbpl_spline_primitives-viz-qt5 base-types trajectory_follower-viz-qt5
     )
