### 2024-11-15 b77bc1d
Finding qt4/qt5 is only required for the GUI
