+++ src/gui/CMakeLists.txt
@@ -1,3 +1,6 @@
+rock_find_qt4(OPTIONAL)
+rock_find_qt5(OPTIONAL)
+
 find_package(PCL 1.7 REQUIRED COMPONENTS common io)
 IF("${PCL_VERSION_MAJOR}.${PCL_VERSION_MINOR}" VERSION_LESS 1.14)
     SET(PCL_VERSION_SUFFIX "-${PCL_VERSION_MAJOR}.${PCL_VERSION_MINOR}")
