### 2024-11-15 4d71537
add vizkit3d and vizkit_debug_drawings check
