+++ src/gui/CMakeLists.txt
@@ -1,5 +1,7 @@
 rock_find_qt4(OPTIONAL)
 rock_find_qt5(OPTIONAL)
+pkg_check_modules(vizkit3d vizkit3d)
+pkg_check_modules(vizkit3d-qt5 vizkit3d-qt5)
 
 find_package(PCL 1.7 REQUIRED COMPONENTS common io)
 IF("${PCL_VERSION_MAJOR}.${PCL_VERSION_MINOR}" VERSION_LESS 1.14)
@@ -7,71 +9,79 @@ IF("${PCL_VERSION_MAJOR}.${PCL_VERSION_MINOR}" VERSION_LESS 1.14)
 ENDIF()
 
 if(ROCK_QT_VERSION_4)
-    rock_library(ugv_nav4d_gui
+    if (vizkit3d_FOUND)
+        rock_library(ugv_nav4d_gui
+            SOURCES
+                PlannerGui.cpp
+            HEADERS
+                PlannerGui.h
+            MOC
+                PlannerGui.h
+            DEPS ugv_nav4d Qt4::QtCore
+            DEPS_PKGCONFIG vizkit3d
+                        vizkit3d-viz
+                        maps-viz
+                        base-viz
+                        sbpl_spline_primitives-viz
+                        base-types
+                        trajectory_follower-viz
+                        pcl_common${PCL_VERSION_SUFFIX}
+                        pcl_io${PCL_VERSION_SUFFIX}
+        )
+
+        rock_executable(ugv_nav4d_bin
         SOURCES
-            PlannerGui.cpp
-        HEADERS
-            PlannerGui.h
-        MOC
-            PlannerGui.h
-        DEPS ugv_nav4d Qt4::QtCore
-        DEPS_PKGCONFIG vizkit3d
-                       vizkit3d-viz
-                       maps-viz
-                       base-viz
-                       sbpl_spline_primitives-viz
-                       base-types
-                       trajectory_follower-viz
-                       pcl_common${PCL_VERSION_SUFFIX}
-                       pcl_io${PCL_VERSION_SUFFIX}
-    )
+                Main.cpp
+        DEPS ugv_nav4d ugv_nav4d_gui
+        )
 
-    rock_executable(ugv_nav4d_bin
-       SOURCES
-            Main.cpp
-       DEPS ugv_nav4d ugv_nav4d_gui
-    )
-
-    rock_executable(ugv_nav4d_replay
-       SOURCES
-            ReplayDump.cpp
-       DEPS ugv_nav4d ugv_nav4d_gui
-       DEPS_PKGCONFIG vizkit3d  maps-viz base-viz
-       sbpl_spline_primitives-viz base-types trajectory_follower-viz
-    )
+        rock_executable(ugv_nav4d_replay
+        SOURCES
+                ReplayDump.cpp
+        DEPS ugv_nav4d ugv_nav4d_gui
+        DEPS_PKGCONFIG vizkit3d  maps-viz base-viz
+        sbpl_spline_primitives-viz base-types trajectory_follower-viz
+        )
+    else()  
+		message(STATUS "vizkit3d not found ... NOT building the ugv_nav4d_gui")
+	endif()
 endif(ROCK_QT_VERSION_4)
 
 if(ROCK_QT_VERSION_5)
-    rock_library(ugv_nav4d_gui-qt5
-        SOURCES
-            PlannerGui.cpp
-        HEADERS
-            PlannerGui.h
-        MOC5
-            PlannerGui.h
-        DEPS ugv_nav4d Qt5::Core
-        DEPS_PKGCONFIG vizkit3d-qt5
-                       vizkit3d-viz-qt5
-                       maps-viz-qt5
-                       base-viz-qt5
[…]
