### 2024-11-15 1c899e2
Improve logging of errors
