+++ src/Planner.cpp
@@ -126,6 +126,7 @@ bool Planner::tryGoal(const Eigen::Vector3d& translation, const double yaw) noex
     }
     catch(const std::runtime_error& ex)
     {
+        LOG_ERROR_S << "Caught exception while setting goal pose:"  << ex.what();
         return false;
     }
     return true;
@@ -138,7 +139,7 @@ Planner::PLANNING_RESULT Planner::plan(const base::Time& maxTime, const base::sa
                                        bool dumpOnError, bool dumpOnSuccess)
 {
 
-    LOG_INFO_S << "Planning with " << plannerConfig.numThreads << " threads";
+    LOG_DEBUG_S << "Planning with " << plannerConfig.numThreads << " threads";
     omp_set_num_threads(plannerConfig.numThreads);
 #ifdef ENABLE_DEBUG_DRAWINGS
     V3DD::CLEAR_DRAWING("ugv_nav4d_successors");
@@ -166,8 +167,8 @@ Planner::PLANNING_RESULT Planner::plan(const base::Time& maxTime, const base::sa
     startbody2Mls.setTransform(startbody2Mls.getTransform());
     endbody2Mls.setTransform(endbody2Mls.getTransform());
 
-    LOG_INFO_S << "start_pose position (raw): " << start_pose.position.transpose();
-    LOG_INFO_S << "end_pose position (raw): " << end_pose.position.transpose();
+    LOG_DEBUG_S << "start_pose position (raw): " << start_pose.position.transpose();
+    LOG_DEBUG_S << "end_pose position (raw): " << end_pose.position.transpose();
 
     const Eigen::Affine3d startGround2Mls(startbody2Mls.getTransform() * ground2Body);
     const Eigen::Affine3d endGround2Mls(endbody2Mls.getTransform() *ground2Body);
@@ -187,13 +188,14 @@ Planner::PLANNING_RESULT Planner::plan(const base::Time& maxTime, const base::sa
     }
     catch(const ugv_nav4d::ObstacleCheckFailed& ex)
     {
-        LOG_INFO_S << "Start inside obstacle.";
+        LOG_ERROR_S << "Caught exception while setting start pose:"  << ex.what();
         if(dumpOnError)
             PlannerDump dump(*this, "start_inside_obstacle", maxTime, startbody2Mls, endbody2Mls);
         return START_INVALID;
     }
     catch(const std::runtime_error& ex)
     {
+        LOG_ERROR_S << "Caught exception while setting start pose:"  << ex.what();
         if(dumpOnError)
             PlannerDump dump(*this, "bad_start", maxTime, startbody2Mls, endbody2Mls);
         return START_INVALID;
@@ -218,7 +220,7 @@ Planner::PLANNING_RESULT Planner::plan(const base::Time& maxTime, const base::sa
     }
     catch(const SBPL_Exception& ex)
     {
-        LOG_ERROR_S << "caught sbpl exception: " << ex.what();
+        LOG_ERROR_S << "Caught SBPL exception: " << ex.what();
         return NO_SOLUTION;
     }
 
@@ -240,39 +242,31 @@ Planner::PLANNING_RESULT Planner::plan(const base::Time& maxTime, const base::sa
 
     try
     {
-        LOG_INFO_S << "Initial Epsilon: " << plannerConfig.initialEpsilon << ", steps: " << plannerConfig.epsilonSteps;
+        LOG_DEBUG_S << "Initial Epsilon: " << plannerConfig.initialEpsilon << ", steps: " << plannerConfig.epsilonSteps;
         planner->set_eps_step(plannerConfig.epsilonSteps);
         planner->set_initialsolution_eps(plannerConfig.initialEpsilon);
 
         solutionIds.clear();
         if(!planner->replan(maxTime.toSeconds(), &solutionIds))
         {
[…]
