+++ src/EnvironmentXYZTheta.cpp
@@ -24,7 +24,7 @@ namespace ugv_nav4d
     {\
         LOG_ERROR_S << #val; \
         LOG_ERROR_S << __FILE__ << ": " << __LINE__; \
-        throw std::runtime_error("meeeeh"); \
+        throw std::runtime_error("Error!"); \
     }
 
 EnvironmentXYZTheta::EnvironmentXYZTheta(std::shared_ptr<MLGrid> mlsGrid,
@@ -103,9 +103,10 @@ void EnvironmentXYZTheta::setInitialPatch(const Eigen::Affine3d& ground2Mls, dou
 
 void EnvironmentXYZTheta::updateMap(shared_ptr< ugv_nav4d::EnvironmentXYZTheta::MLGrid > mlsGrid)
 {
-    if(this->mlsGrid && this->mlsGrid->getResolution() != mlsGrid->getResolution())
+    if(this->mlsGrid && this->mlsGrid->getResolution() != mlsGrid->getResolution()){
+        LOG_ERROR_S << "EnvironmentXYZTheta::updateMap : Error got MLSMap with different resolution";
         throw std::runtime_error("EnvironmentXYZTheta::updateMap : Error got MLSMap with different resolution");
-
+    }
     if(!this->mlsGrid)
     {
         availableMotions.computeMotions(mlsGrid->getResolution().x(), travConf.gridResolution);
@@ -131,7 +132,7 @@ EnvironmentXYZTheta::ThetaNode* EnvironmentXYZTheta::createNewStateFromPose(cons
     traversability_generator3d::TravGenNode *travNode = travGen.generateStartNode(pos);
     if(!travNode)
     {
-        LOG_INFO_S << "Could not generate Node at pos";
+        LOG_ERROR_S << "Could not generate new state at pos: " << pos.transpose();
         return nullptr;
     }
 
@@ -140,7 +141,7 @@ EnvironmentXYZTheta::ThetaNode* EnvironmentXYZTheta::createNewStateFromPose(cons
     {
         if(!travGen.expandNode(travNode))
         {
-            LOG_INFO_S << "createNewStateFromPose: Error: " << name << " Pose " << pos.transpose() << " is not traversable";
+            LOG_ERROR_S << "createNewStateFromPose: Error: " << name << " Pose " << pos.transpose() << " is not traversable";
             return nullptr;
         }
         travNode->setNotExpanded();
@@ -165,13 +166,13 @@ bool EnvironmentXYZTheta::obstacleCheck(const maps::grid::Vector3d& pos, double 
     maps::grid::Index idxObstNode;
     if(!obsGen.getTraversabilityMap().toGrid(pos, idxObstNode))
     {
-        LOG_INFO_S << "Error " << nodeName << " is outside of obstacle map ";
+        LOG_ERROR_S << "Error " << nodeName << " is outside of obstacle map ";
         return false;
     }
     traversability_generator3d::TravGenNode* obstacleNode = obsGen.findMatchingTraversabilityPatchAt(idxObstNode, pos.z());
     if(!obstacleNode)
     {
-        LOG_INFO_S << "Error, could not find matching obstacle node for " << nodeName;
+        LOG_ERROR_S << "Error, could not find matching obstacle node for " << nodeName;
         return false;
     }
 
@@ -209,8 +210,8 @@ bool EnvironmentXYZTheta::obstacleCheck(const maps::grid::Vector3d& pos, double 
             });
 #endif
 
-            LOG_INFO_S << "Num obstacles: " << stats.getRobotStats().getNumObstacles();
-            LOG_INFO_S << "Error: " << nodeName << " inside obstacle";
[…]
