### 2024-11-18 76ebeef
Fix: Check for all of vizkit3d, vizkit3d_debug_drawings libraries
The debug drawings are marked optional, so we must check for all of them,
vizkit3d-qt5 and vizkit3d(qt4 version) can be absent depending on the
version of vizkit3d being compiled or settings in vizkit3d, independent
of the presence of Qt4 or Qt5.
