### 2024-11-18 03ac049
Need to install dependencies before using Rock macros
