### 2024-11-19 #6983 @soraxas
@logological  I'm happy to acceptance this with the unit test that are to be added in PR https://github.com/dfki-ric/ugv_nav4d/pull/18
