### 2024-11-19 dfki-ric/ugv_nav4d#8 @soraxas
Good to close this when #18 is merged
