### 2024-11-19 8af2e4c
Add unit tests (#18)
* Add unit tests for environmentxyztheta

* Add boost::filesystem library (must be after ugv_nav4d)

* Add unit tests

* Cleanup headers

* Use set locale to C for grid resolution parameter from cli for gui

* Throw error if translation or angular velocity is 0.0

* Add unit tests for computed motions

* Update readme for unit tests

* Test costs calculation for only one motion

* Check if final pose of full and sampled splines

* Check validity of underlying values in generated trajectories

* Use the ugv_nav4d target for tests

* Use rock executable macro for tests

* Ad
[…]
