+++ src/test/test_ugv_nav4d.cpp
@@ -1,274 +1,0 @@
-#define BOOST_TEST_MODULE PlannerTestModule
-#include <boost/test/included/unit_test.hpp>
-
-#include <fstream>
-#include <cstdlib>
-
-#include "ugv_nav4d/DiscreteTheta.hpp"
-#include "ugv_nav4d/Planner.hpp"
-#include <traversability_generator3d/TraversabilityConfig.hpp>
-#include <maps/grid/MLSMap.hpp>
-
-#include <pcl/io/ply_io.h>
-#include <pcl/common/common.h>
-#include <pcl/common/transforms.h>
-
-using namespace ugv_nav4d;
-
-std::string filePath;
-
-class PlannerTest {
-public:
-    PlannerTest();
-    ~PlannerTest();
-    void loadMlsMap(const std::string& path);
-    std::string getResult(const Planner::PLANNING_RESULT& result);
-
-    Planner* planner;
-    maps::grid::MLSMapSloped mlsMap;
-    PlannerConfig plannerConfig;
-    Mobility mobility;
-    sbpl_spline_primitives::SplinePrimitivesConfig splinePrimitiveConfig;
-    traversability_generator3d::TraversabilityConfig traversabilityConfig;
-    bool map_loaded = false;
-};
-
-PlannerTest::PlannerTest() {
-    splinePrimitiveConfig.gridSize = 0.3;
-    splinePrimitiveConfig.numAngles = 42;
-    splinePrimitiveConfig.numEndAngles = 21;
-    splinePrimitiveConfig.destinationCircleRadius = 10;
-    splinePrimitiveConfig.cellSkipFactor = 3;
-    splinePrimitiveConfig.splineOrder = 4.0;
-
-    mobility.translationSpeed = 0.5;
-    mobility.rotationSpeed = 0.5;
-    mobility.minTurningRadius = 1;
-    mobility.spline_sampling_resolution = 0.05;
-    mobility.remove_goal_offset = true;
-    mobility.multiplierForward = 1;
-    mobility.multiplierBackward = 3;
-    mobility.multiplierPointTurn = 3;
-    mobility.multiplierLateral = 4;
-    mobility.multiplierForwardTurn = 2;
-    mobility.multiplierBackwardTurn = 4;
-    mobility.multiplierLateralCurve = 4;
-    mobility.searchRadius = 0.0;
-    mobility.searchProgressSteps = 0.1;
-    mobility.maxMotionCurveLength = 100;
-
-    traversabilityConfig.maxStepHeight = 0.25;
-    traversabilityConfig.maxSlope = 0.45;
-    traversabilityConfig.inclineLimittingMinSlope = 0.2;
-    traversabilityConfig.inclineLimittingLimit = 0.1;
-    traversabilityConfig.costFunctionDist = 0.0;
-    traversabilityConfig.minTraversablePercentage = 0.4;
-    traversabilityConfig.robotHeight = 1.2;
-    traversabilityConfig.robotSizeX = 1.35;
-    traversabilityConfig.robotSizeY = 0.85;
-    traversabilityConfig.distToGround = 0.0;
-    traversabilityConfig.slopeMetricScale = 1.0;
-    traversabilityConfig.slopeMetric = traversability_generator3d::NONE;
-    traversabilityConfig.gridResolution = 0.3;
-    traversabilityConfig.initialPatchVariance = 0.0001;
-    traversabilityConfig.allowForwardDownhill = true;
-    traversabilityConfig.enableInclineLimitting = false;
-
-    plannerConfig.searchUntilFirstSolution = false;
-    plannerConfig.initialEpsilon = 64;
-    plannerConfig.epsilonSteps = 2;
[…]
