+++ src/test/test_Planner.cpp
@@ -1,0 +1,264 @@
+#define BOOST_TEST_MODULE PlannerTestModule
+#include <boost/test/included/unit_test.hpp>
+
+#include "ugv_nav4d/Planner.hpp"
+#include <base/Angle.hpp>
+
+using namespace ugv_nav4d;
+
+class PlannerTest {
+public:
+    PlannerTest();
+    ~PlannerTest();
+    std::string getResult(const Planner::PLANNING_RESULT& result);
+
+    Planner* planner;
+    maps::grid::MLSMapSloped* mlsMap;
+    PlannerConfig plannerConfig;
+    Mobility mobility;
+    sbpl_spline_primitives::SplinePrimitivesConfig splinePrimitiveConfig;
+    traversability_generator3d::TraversabilityConfig traversabilityConfig;
+};
+
+PlannerTest::PlannerTest() {
+
+    splinePrimitiveConfig.gridSize = 0.3;
+    splinePrimitiveConfig.numAngles = 16;
+    splinePrimitiveConfig.numEndAngles = 8;
+    splinePrimitiveConfig.destinationCircleRadius = 6;
+    splinePrimitiveConfig.cellSkipFactor = 0.1;
+    splinePrimitiveConfig.splineOrder = 4.0;
+
+    mobility.translationSpeed = 0.5;
+    mobility.rotationSpeed = 0.5;
+    mobility.minTurningRadius = 1;
+    mobility.spline_sampling_resolution = 0.05;
+    mobility.remove_goal_offset = true;
+    mobility.multiplierForward = 1;
+    mobility.multiplierBackward = 2;
+    mobility.multiplierPointTurn = 3;
+    mobility.multiplierLateral = 4;
+    mobility.multiplierForwardTurn = 2;
+    mobility.multiplierBackwardTurn = 3;
+    mobility.multiplierLateralCurve = 4;
+    mobility.searchRadius = 1.0;
+    mobility.searchProgressSteps = 0.1;
+    mobility.maxMotionCurveLength = 100;
+
+    traversabilityConfig.maxStepHeight = 0.2;
+    traversabilityConfig.maxSlope = 0.45;
+    traversabilityConfig.inclineLimittingMinSlope = 0.2;
+    traversabilityConfig.inclineLimittingLimit = 0.1;
+    traversabilityConfig.costFunctionDist = 0.0;
+    traversabilityConfig.minTraversablePercentage = 0.4;
+    traversabilityConfig.robotHeight = 0.5;
+    traversabilityConfig.robotSizeX = 0.5;
+    traversabilityConfig.robotSizeY = 0.5;
+    traversabilityConfig.distToGround = 0.0;
+    traversabilityConfig.slopeMetricScale = 1.0;
+    traversabilityConfig.slopeMetric = traversability_generator3d::NONE;
+    traversabilityConfig.gridResolution = 0.3;
+    traversabilityConfig.initialPatchVariance = 0.0001;
+    traversabilityConfig.allowForwardDownhill = true;
+    traversabilityConfig.enableInclineLimitting = false;
+
+    plannerConfig.searchUntilFirstSolution = false;
+    plannerConfig.initialEpsilon = 64;
+    plannerConfig.epsilonSteps = 2;
+    plannerConfig.numThreads = 4;
+
+    maps::grid::Vector2d res(0.3, 0.3);
+    maps::grid::Vector2ui numCells(100, 100);
+
+    maps::grid::MLSConfig mls_config;
+    mls_config.gapSize = 0.1;
+    mls_config.updateModel = maps::grid::MLSConfig::SLOPE;
+    
+    mlsMap = new maps::grid::MLSMapSloped(numCells, res, mls_config);
+
+    /** Translate the local frame (offset) **/
[…]
