### 2024-11-19 dfki-ric/ugv_nav4d#8 @haider8645
Thank you! closing this.
