### 2024-11-20 #6983 @logological
@haider8645  I've noticed a few minor problems in your bibliography; could you please address these?

* The references are inconsistent as to whether URLs are hyperlinked.  Personally I think hyperlinking the URLs is helpful to the reader; could you please make the URLs for Born & Arnold (2015) and Machowinski & Böckmann (2016) hyperlinks?
* In the title of the Macenski et al. (2022) reference, "Robot Operating System" should be in title case, since it's a proper name.  Could you please ensure all three words are capitalized?
* The Robotics Exploration Laboratory reference has the same issue – please use title case for this proper name.
* There are a couple entries in your `.bib` file that are not cited in the paper.  Though it's not strictly necessary to do so, you might consider removing these, if only to spare the editor-in-chief the trouble of tracking down spurious warnings from our automated reference checker.
