### 2024-11-20 #6983 @logological
@haider8645 Some further minor issues with the paper that you may want to address:

* At least one of the authors (namely, yourself) appears to have an ORCID.  Though it's not a requirement, you might consider adding at least your own ORCID to the paper for [the various benefits this may bring](https://info.orcid.org/researchers/#why).
* When citing several references at once, as in the first paragraph of your paper, consider combining them into a single citation command.  That is, you can write `[@mlsmaps; @slammaps]` instead of `[@mlsmaps], [@slammaps]`.  This will format the citations in a more standard style.
* Your section headings are inconsistent as to whether they use sentence case or title case.  (In particular, you have "Statement of need" in sentence case but "Software Components" and "Field Tests" in title case.)  I don't think we mandate one style or the other, but you should at least pick one style and use it consistently.
* The paper seems to be written in Commonwealth (British) English, except for the word "modeling" which is in American English.  Could I suggest changing the spelling to "modelling" for consistency with the rest of the paper?
