+++ paper/paper.bib
@@ -1,61 +1,58 @@
 @software{slammaps,
+  title = {{SLAM Maps}},
   author = {Anna Born and Sascha Arnold},
-  title = {Slam Maps},
   year = {2015},
-  howpublished = {https://github.com/envire/slam-maps},
+  url = {https://github.com/envire/slam-maps},
+  journal = {GitHub repository}
 }
 
 @software{sbpl,
+  title        = {{SBPL: Search-Based Planning Library}},
   author       = {Maxim Likhachev and Dave Ferguson and Geoff Gordon},
-  title        = {{SBPL}: Search-Based Planning Library},
   year         = {2005},
   version      = {1.3},
-  url          = {https://github.com/sbpl/sbpl}
+  url          = {https://github.com/sbpl/sbpl},
+  journal      = {GitHub repository}
 }
 
 @software{travgen3d,
+  title = {{Traversability Generator3d}},
   author = {Janosch Machowinski and Arne Böckmann},
-  title = {Traversability Generator3d},
-  howpublished = {https://github.com/dfki-ric/traversability_generator3d.git},
+  url = {https://github.com/dfki-ric/traversability_generator3d},
   note = {Accessed: June 20, 2024},
-  year = {2016}
+  year = {2016},
+  journal = {GitHub repository}
 }
 
 @misc{roboticsElab,
-  title = {Robotics Exploration Laboratory},
+  title = {{Robotics Exploration Laboratory}},
   url = {https://www.dfki.de/web/technologien-anwendungen/living-labs/robotics-exploration-laboratory},
   note = {Accessed: June 20, 2024},
   year = {2024}
 }
 
 @software{trajectoryfollower,
+  title = {{Trajectory Follower}},
   author = {Ajish Babu},
-  title = {Trajectory Follower},
   url = {https://github.com/rock-control/control-trajectory_follower},
   note = {Accessed: June 20, 2024},
-  year = {2016}
+  year = {2016},
+  journal = {GitHub repository}
 }
 
 @software{sbplsplineprimitives,
+  title = {{SBPL Spline Primitives}},
   author = {Arne Böckmann},
-  title = {SBPL Spline Primitives},
   url = {https://github.com/rock-planning/planning-sbpl_spline_primitives},
   note = {Accessed: June 20, 2024},
-  year = {2016}
-}
-
-@software{googletest,
-  author = {Google},
-  title = {GoogleTest},
-  url = {https://github.com/google/googletest},
-  note = {Accessed: June 20, 2024},
-  year = {2024}
+  year = {2016},
+  journal = {GitHub repository}
 }
 
 @INPROCEEDINGS{mlsmaps,
+  title={{Multi-Level Surface Maps for Outdoor Terrain Mapping and Loop Closing}}, 
   author={Triebel, Rudolph and Pfaff, Patrick and Burgard, Wolfram},
   booktitle={2006 IEEE/RSJ International Conference on Intelligent Robots and Systems}, 
-  title={Multi-Level Surface Maps for Outdoor Terrain Mapping and Loop Closing}, 
   year={2006},
   volume={},
   number={},
@@ -65,31 +62,19 @@
 }
 
 @article{macenski2020nav2,
+  title        = {{The Marathon 2: {A} Navigation System}},
   author       = {Steve Macenski and
                   Francisco Mart{\'{\i}}n and
                   Ruffin White and
                   Jonatan Gin{\'{e}}s Clavero},
-  title        = {The Marathon 2: {A} Navigation System},
   journal      = {CoRR},
   volume       = {abs/2003.00368},
   year         = {2020},
   url          = {https://arxiv.org/abs/2003.00368},
-  eprinttype    = {arXiv},
[…]
