### 2024-11-21 82bb3c0
Minor fixes to remove most warnings
