### 2024-11-21 #6983 @logological
@haider8645 Thanks for making those changes.  However, I see you made a few additional changes that may have introduced some new issues:

1. Added a new field `journal = {GitHub repository}` to the software references.  I don't think that this is helpful, particularly since our bibliography style words and formats the reference as though "GitHub repository" is the name of a titled work in which the software can be found: "… Trajectory Follower. In _GitHub repository_. …"  I suggest you remove these fields.
2. You have forced title case for the title of the Triebel et al. (2006) reference, which overrides our default style.  Could you please remove the extra pair of braces here to let our bibliography tool format the title in the house style?
3. You replaced the citation to Macenski et al. (2022) with a citation to a different paper, Macenski et al. (2020).  Can you confirm that this was this intentional?  (The new paper seems like a better fit to me, but I just wanted to double-check with you that this is what you wanted.)
