### 2024-11-22 5f5d505
Update paper bib
