+++ paper/paper.bib
@@ -3,7 +3,6 @@
   author = {Anna Born and Sascha Arnold},
   year = {2015},
   url = {https://github.com/envire/slam-maps},
-  journal = {GitHub repository}
 }
 
 @software{sbpl,
@@ -12,7 +11,6 @@
   year         = {2005},
   version      = {1.3},
   url          = {https://github.com/sbpl/sbpl},
-  journal      = {GitHub repository}
 }
 
 @software{travgen3d,
@@ -21,7 +19,6 @@
   url = {https://github.com/dfki-ric/traversability_generator3d},
   note = {Accessed: June 20, 2024},
   year = {2016},
-  journal = {GitHub repository}
 }
 
 @misc{roboticsElab,
@@ -37,7 +34,6 @@
   url = {https://github.com/rock-control/control-trajectory_follower},
   note = {Accessed: June 20, 2024},
   year = {2016},
-  journal = {GitHub repository}
 }
 
 @software{sbplsplineprimitives,
@@ -46,16 +42,13 @@
   url = {https://github.com/rock-planning/planning-sbpl_spline_primitives},
   note = {Accessed: June 20, 2024},
   year = {2016},
-  journal = {GitHub repository}
 }
 
 @INPROCEEDINGS{mlsmaps,
-  title={{Multi-Level Surface Maps for Outdoor Terrain Mapping and Loop Closing}}, 
+  title={Multi-Level Surface Maps for Outdoor Terrain Mapping and Loop Closing}, 
   author={Triebel, Rudolph and Pfaff, Patrick and Burgard, Wolfram},
   booktitle={2006 IEEE/RSJ International Conference on Intelligent Robots and Systems}, 
   year={2006},
-  volume={},
-  number={},
   pages={2276-2282},
   keywords={Terrain mapping;Mobile robots;Data structures;Path planning;Multilevel systems;Bridges;Buildings;Impedance matching;Robot sensing systems;Large-scale systems},
   doi={10.1109/IROS.2006.282632}
