### 2024-11-22 #6983 @haider8645
@logological I violated the `never change a working system..` rule. Implemented your suggested changes.

about the `journal = {GitHub repository}`: I felt it is unnecessary as well but used it after finding it in the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html#:~:text=journal%20%3D%20%7BGitHub%20repository%7D%2C)
