### 2024-11-25 b3fba76
Fix for the s-shaped 3d trajectory
