### 2024-11-25 e087fe9
Fix for the s-shaped 3d trajectory
