### 2024-11-25 703315f
Remove extra variable
