### 2024-11-25 #6983 @logological
@haider8645 Thanks; everything looks good now.  Could you please now handle the author tasks listed in [my earlier comment](https://github.com/openjournals/joss-reviews/issues/6983#issuecomment-2492440178)?  Let me know once these are complete so that I can do a final check.
