### 2024-11-26 8c3b52d
Add next pose to the previous global pose in the path
