### 2024-11-26 02977cd
Get intersection of point with plane along +z-axis
