+++ src/EnvironmentXYZTheta.hpp
@@ -200,7 +200,8 @@ public:
     std::vector<Motion> getMotions(const std::vector<int> &stateIDPath);
 
     void getTrajectory(const std::vector<int> &stateIDPath, std::vector<trajectory_follower::SubTrajectory> &result,
-                       bool setZToZero, const Eigen::Vector3d &startPos, const Eigen::Vector3d &goalPos, const double& goalHeading, const Eigen::Affine3d &plan2Body = Eigen::Affine3d::Identity());
+                       const Eigen::Vector3d &startPos, const Eigen::Vector3d &goalPos, const double& goalHeading, 
+                       const Eigen::Affine3d &plan2Body = Eigen::Affine3d::Identity());
 
     const PreComputedMotions& getAvailableMotions() const;
 
