### 2024-11-26 690d349
Add option to generate trajectory2D for backward compatibility
