### 2024-11-26 #6983 @haider8645
> * Archive the release on Zenodo/figshare/etc and post the DOI here.
1) Done
2) Release version: v1.0.0
3) Zenodo: https://doi.org/10.5281/zenodo.14224714
4) Done
5) Done
