### 2024-11-26 #6983 @logological
@haider8645 Thanks.  It seems that there are some metadata discrepancies between the submitted paper and the Zenodo archive:

1. The titles are different.  In the paper it is "Advanced Multi-Surface Navigation for Unmanned Ground Vehicles (UGVs) Using 4D Path Planning Techniques (ugv_nav4d)" but in the Zenodo archive it is "Advanced Multi-Surface Navigation for Unmanned 2 Ground Vehicles (UGVs) Using 4D Path Planning 3 Techniques (ugv_nav4d)".  I assume the extraneous "2" and "3" are copy-paste errors.  Could you please fix these?
2. The order of the authors is different.  In the paper it is Böckmann, Machowinski, Lodhi, but in the Zenodo archive it is Lodhi, Böckmann, Machowinski.  Did Zenodo give you the option of specifying the order of the authors?  If so, please change it to match the order in the paper.
