### 2024-11-27 #6983 @crvernon
:wave:  @haider8645 - I just need you to address the following before I move to accept this for publication:

- [ ] I noticed the version in your GitHub repository is now v1.0.1 - this is not in agreement with what is listed in your Zenodo archive and what has been declared in your review thread.  Do you want to change to final version to v1.0.1?

In the paper:

- [ ]  In the title, please remove what you have in"(ugv_nav4d)" at the end.  If you want the name of the package in the title, may I recommend:  "ugv_nav4d:  Advanced Multi-Surface Navigation for Unmanned Ground Vehicles using 4D Path Planning Techniques". Any changes you make to the title, please also make to the title in your Zenodo archive by simply editing the metadata.  No new release necessary.
- [ ] LINE 39: "meta data" should be one word "metadata"

Let me know when you have addressed these and I'll move to accept.  Thanks!
