### 2024-11-28 #6983 @haider8645
Hello @crvernon,

thanks for the feedback. I have updated the paper and zenodo based on your suggestions. v1.0.0 is fixed to the last commit related to the review. Afterwards, I added new commits (some nice-to-have enhancements). Therefore, I kept version v1.0.0 in zenodo because it would be linked to the paper. If it is permissible from your side then I can update the version to v1.0.1 in zenodo as well.
