### 2024-11-28 1fcbc6f
Add gui for visualization of spline primitives
