+++ README.md
@@ -97,6 +97,9 @@ The button `Create PlannerDump` can be used to save the planner's state. The cre
 ugv_nav4d_replay ugv4d_dump_xxxx.bin
 ```
 
+An additional GUI is provided for tuning of parameters used in the generation of motion primitives. The generated spline motion primitives are also visualized. Furthermore, the `vizkit3d::SbplSplineVisualization` plugin under `Properties` offers further options for visualizing and analyzing the splines for various start and end angles.
+![MotionPrimitivesGui](doc/figures/motion_primitives_gui.png)
+
 #### Unit Tests
 
 Build the library again but this time enable the `-DTESTS_ENABLED=ON`
