### 2024-11-28 152ff22
Add command for spline visualization gui
