### 2024-11-28 #6983 @crvernon
@haider8645 yes it is fine to switch to version v1.0.1
I'll go ahead and do so here.  You will just need to update the Zenodo archive.  Let me know when this has been done.
