+++ src/Planner.hpp
@@ -135,7 +135,10 @@ public:
     const maps::grid::TraversabilityMap3d<traversability_generator3d::SoilNode*> &getSoilMap() const;
 
     const maps::grid::TraversabilityMap3d<traversability_generator3d::TravGenNode*> &getObstacleMap() const;
-    
+
+    void clearSoilMap();
+    void addSoilNode(const traversability_generator3d::SoilSample& sample);
+
     std::shared_ptr<trajectory_follower::SubTrajectory> findTrajectoryOutOfObstacle(const Eigen::Vector3d& start, double theta,
             const Eigen::Affine3d& ground2Body);
 
