### 2024-12-02 #6983 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
